[roll] Roll fuchsia [vfs] Remove Journal OperationTracker
Replace the Journal OperationTracker with the equivalent data structure.

Discovered as part of metrics cleanup, removed to reduce confusion with
other metrics/statistics-related trackers. The only functionality that
OperationTracker provided which the backing IntervalTree does not is
the Remove function, which was not used in production (only tests).

Remove can be implemented as a free function if required in the future.

Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683629
Original-Revision: 4c0ceba693f17a976988a08633312296046d6827
GitOrigin-RevId: dc7cc1331607a29e0aa6e7f1f367e2c1c63897af
Change-Id: I214887da04b52e4a7ef93af3bc806066c4d4797c