[roll] Roll fuchsia [kernel][vm] Partial failures for Internal batch requests
All DIRTY requests operate in internal batch mode, where the PageSource
accumulates multiple pages into the same request instead of sending per
page requests. This is an optimization, so failures that do not involve
the first page should not be treated as fatal for the request in its
entirety. This differs from the behavior for externally batched requests
(currently only used by commit) where a failure in any portion of the
request is treated as fatal and must be propagated to the user.

Formalize the notion of Internal batches by using the PageRequest
batch_state_. For partial failures that do not touch the first page of
Internal requests, ignore the failure and let the external caller retry.
The retry approach also fits in with the existing model of page request
failures, where we do not remember failure state for pages, i.e. the
failure state is not sticky and is solely used for signaling. This
change allows the caller to operate on any pages that came before the
failing offset if they had been successfully resolved.

This addresses an inconsistency in VMO write where if a single DIRTY
page request only partially succeeded, we would fail the entire page
request and not copy in data to the portion before the failing offset
even though those pages had successfully been marked dirty by the user
pager. So the caller would observe the write partially succeeding until
an earlier offset compared to what the userpager indicated. This makes
the behavior consistent with partial writes where if page requests fail
as a whole, all data until the failing request will have been
successfully copied in.

Note that this case is only relevant for streams which have a notion of
partial success for writes; zx_vmo_write succeeds only if the entire
range succeeds and can tolerate an undefined number of bytes being
copied in case of failure.

Test: core-pager-writeback
Original-Bug: 63989
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684370
Original-Revision: 28a151fc847538fd12a971b18bba0e42ffaa1751
GitOrigin-RevId: bb1a892e435c17e4415812506d0a137bb618d856
Change-Id: I94b6f25c0ce74043d31291f63fcf46ef40100c75