[roll] Roll fuchsia [scripts/owner] Find all target/files references
Extend `suggest_owners.py` to accept an `--all_refs` argument to search
for references to all targets (not just the top-level targets), and to
all files in the project if no target references were found.

This is useful for finding owners for projects that either don't have a
top-level gn target, or are not built with gn at all and are referenced
directly by file instead.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684562
Original-Revision: ffb453138ba1383e4b208038c566823ee1359d99
GitOrigin-RevId: a73612d0fdfa7607a3a2d86a8d94de75d15a386e
Change-Id: I2d3c071477e775164e33b26e9afad0c5901c5dd3