[roll] Roll llvm-project/libc [libc] add fprintf and file_writer
This patch adds the file_writer header, which just provides a wrapper
for File->write, as well as fprintf to use it. There are no unit tests
for file_writer since it's too simple to need them, but fprintf does
have a simple test of writing to a file.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D125939

GitOrigin-RevId: 9bb23ae2b55e06b11cc4c0adc42b2e19e179687b
Original-Revision: 46796416d7479ca1813c6a041caa33f0dd6173ab
Change-Id: Ia025a9d69bb62e291f740cb5102e0e2a0e330d6b