[roll] Roll fuchsia [minfs] separate the vfs object from the minfs object
Minfs and the vfs are closely linked, but we need to be able to start
one without having the information we need to start the other in order
to componentize the filesystems. This breaks the vfs object out into a
Runner class, which owns the Minfs object once it's created.

Right now, the runner class just immediately creates the minfs object.
There are a couple of extra accessors over the version of the Runner
class that is in blobfs because minfs tests love running the minfs code
directly.

Original-Fixed: 100846
Test: fx test //src/storage/minfs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683138
Original-Revision: 7720898169127b370996ccfdcad6e3ce84f824e3
GitOrigin-RevId: c28e7d855000dbdbb996578caf7337c862595e6a
Change-Id: Ifbc15ed2418721682fa6521721146ab8ccdd3ec4