[roll] Roll fuchsia [component framework] Improve lifecycle unresolve() errors
For lifecycle unresolve(), return component-not-found and
unable-to-unresolve errors.

The `ffx component reload` command first executes a ShutdownAction,
which would error if applied to a nonexistent component, so this
issue is not user-facing. However, it is updated here for
correctness.

Original-Bug: 101330
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685282
Original-Revision: 086fbd8c011a5da066b606d54671d1673dd2281f
GitOrigin-RevId: 7e2e55bdc2624a2174616927c9a0b9cfa77e9444
Change-Id: I6974f4adbd7d28604c394565c8fad686159d29dd