[roll] Roll fuchsia [ffx][config] Don't call config.save() twice.
As-is, the 'else' path of this function will always be called, even
if there's a build_dir set. To make it not call it, map_or_else
would need to be used. But I think it's clearer to just only
initialize the build_dir and then use it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685324
Original-Revision: 6b42b6004484e8432aaecf1a943fe18163ec953a
GitOrigin-RevId: d7740fd2ba04167a84d472f13b3ebf9d6daa437b
Change-Id: I121ad8e93acc53b3a951c9e9c0c536c134d83c94