[roll] Roll fuchsia [component][fuzzing] Fix mutation_depth
This CL addresses an issue in which the number of mutations was not
being incremented by realm fuzzing engine.

In the process of fixing that issue, this CL revealed a second bug:
seed corpus elements were not being included in the live corpus and
therefore never picked for mutation. This CL also addresses that issue,
and includes a new method and unit tests for adding one corpus to
another.

Finally, this CL tweaks the libFuzzer runner tests in order to make
coverage more predictable in testing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681144
Original-Revision: 238e4e228a9663b0b1bc1573c88d10eee943e2e7
GitOrigin-RevId: 78dd776f7e3c1909bc3e69accb4b963540703ad8
Change-Id: Ie77e200268af24c60ff751dc030f0ee4b44755ee