[roll] Roll fuchsia [assembly] Configure example based on build type.
After discussion it seems like it might be overkill for the example
to have its own platform config, so we'll infer it from the build type
as a demontration for the more common case where people will be doing
that.

Deletes two tests which checked to make sure that the combination of
platform config and experimental ffx flag worked together correctly.
Now the ffx flag is the only thing which determines whether the example
is configured.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685576
Original-Revision: d76908c2fb356c684ae66f8e532b0403dc84e9d7
GitOrigin-RevId: 7ca494d90272ff17c97b9d669aff85e0687e2593
Change-Id: I179176c0c31c300d77b674fb070fbc516b9c93ea