[roll] Roll fuchsia [fxfs] Change the storage benchmarks to use v2 components
This changes the storage benchmarks so that it runs as a v2 component
which allows us to run Fxfs as a component.  To make things simpler,
this change now runs ODU as a library rather than spawning ODU as
another component.  For now, ODU is the only benchmark that we support.

Original-Bug: 99591
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684846
Original-Revision: c0bbab53ec6bef966fd952ac3c0b23fcf60536f1
GitOrigin-RevId: 466d9f296973b01c852651db32dd7152f3d7f697
Change-Id: Idcc43738e766d3366c739d3065201a90b9694dc2