[roll] Roll fuchsia [fidl] Remove debug assertion that boolean value is 0 or 1
There currently is a debug assertion in the bool WireCodingTraits that
ensures that bool values are 0 or 1 in the encode direction. However,
the reinterpret_cast in this check uses UB.

This check should not be needed unless there is memory corruption
because the bool value is already guaranteed to be 0 or 1. Given this,
this CL removes the unnecessary check.

Original-Fixed: 100689
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683679
Original-Revision: 86d24b80b7862dbf71a2528ceab13d58db959b43
GitOrigin-RevId: b24cd949ad33509fcdecd0e2a2ca3ca329e18dad
Change-Id: Iaa936a7f6fe5812a5b7f848fcf01599d5a32e406