[roll] Roll fuchsia [fidlc] Add tests for interleaving library availabilities
This CL extends availability_interleaving_tests.cc to test
availabilities on libraries themselves. It also adds a test for the
@available(platform="...") syntax, since the main change incidentally
reduced coverage of that feature.

In particular, the tests reinforce that the following is allowed:

    // foo.fidl
    @available(added=1)
    library foo;
    using bar;
    @available(added=2)
    alias S = bar.S;

    // bar.fidl
    @available(added=2)
    library bar;
    type S = struct {};

It is not an error for 'foo' to import 'bar', even though 'bar' does not
exist at version 1, because 'foo' does not actually use any element of
'bar' until version 2. This is a consequence of treating absent
libraries as empty (I42f528623bccc38385342b7b889d495e242dbb61).

Test: fx test fidl-compiler
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686110
Original-Revision: 4ab9123f848695d3511b79b6bff9c4d322682d31
GitOrigin-RevId: b50e85828a3f2ef994249b34be45c84924984a72
Change-Id: I5946a125416baa2595aaaa8182b48dc21283c001