[roll] Roll fuchsia [starnix] Allows access to the filesystem withtout a task
Until now, going through the filesystem required having access to a
Task.

It was needed for:
- Resolving symbolic links to absolute path
- Resolving /proc special symbolic links

This prevented accessing the filesystem while working at the runner
level.

This CL move the state state inside the LookupContext and allows it to
be empty, so that special context can be built and used when traversing
the file system outside of the context of a task.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686223
Original-Revision: e016b22d96a42e5437f84f9fad34f9b547d69fa5
GitOrigin-RevId: 350e07eea430bec2f3f7876c9b97750897481fdc
Change-Id: I665527987de5ec2292e26942018d51e6fb480d3a