[roll] Roll fuchsia [ffx] Add console sockets to the flags template
In the attached bug, we have a request to enable the user/ffx emu tool
to communicate with the emulator backend through the use of Linux
Sockets. The QEMU backend supports attaching to sockets for several
devices; most importantly for this work are the QEMU Monitor console,
the QMP (a json-based command processor), and the serial port (which
exposes the Fuchsia console).

This CL adds the appropriate flags to the template file for attaching
sockets to these consoles. Tools to utilize the sockets will be
developed in later CLs.

Original-Bug: 94232
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682465
Original-Revision: 667c4f48a5aa9295ac0c2ba045243240d2dc67bd
GitOrigin-RevId: a15eecfe14ed0ad18cc2212d59b1689e4501265d
Change-Id: Ib9be9b00f67c0afb7336228ff85a84306edc290b