[roll] Roll fuchsia [kernel][benchmarks] Don't trigger lockup detector during `k bench`
When running in-kernel benchmarks (`k bench`) that disable interrupts
for extended periods, mark the CPU as inactive so that the lockup
detector knows that it's OK for it to not update it's heartbeats.  The
lockup detector only monitors active CPUs.

Test: Tested by running `k bench` on QEMU and observing no OOPS.
Original-Fixed: 101553
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686322
Original-Revision: 46ffbc795d819c5b5b02e7777a157e55cf6ac052
GitOrigin-RevId: 2473c2879d1fd27ddea743575a219b208de6debd
Change-Id: Iff9b9f548c69fde7280cba44f6a53c4ac4bade45