[roll] Roll fuchsia [kernel][scheduler] Introduce PreemptDisabled token.
Introduce a token which can be used along with clang static analysis
to require that preemption be disabled for certain sections of code.

Typically, disabling preemption is just an optimization and does not
need this level of enforcement.  There are places, however, where
correctness may demand that preemption be disabled.  In particular,
owned wait queue code demands that preemption be disabled during
priority propagation in order to satisfy its recursion detector
safety.

OWQ interactions already demand that the thread lock be held, so on
the surface it seems like there should be no danger of re-entering
code which is running while the thread lock is held.  This is,
unfortunately, not 100% true.  If a thread is holding the thread lock
(and just the thread lock) when it attempts to interact with the
scheduler in such a way which changes a thread's eligibility to run
(such changes to the PI graph it is involved in), it could result in
an immediate context switch.  In such a case, the thread which was
holding the thread lock would block, and the thread which was suddenly
more runnable would start to execute, holding the thread lock as it
did when it blocked.

The best way to prevent this is to disable local involuntary
preemption while thread lock is held and interacting with the
scheduler.  If any actions the thread takes would have resulted in
involuntary preemption, the preemption will be deferred until after
the thread has dropped the thread lock and re-enabled preemption.

The PreemptDisabled token helps to enforce this requirement in a
fashion similar to how mutex invariants are currently statically
enforced in Zircon, using clang static thread analysis.

The token class itself does very little at runtime, consisting of only
3 methods, Acquire, Release, and AssertHeld, and no member variables.
Acquire and Release do nothing at all at runtime and exist simply to
carry the TA_ACQ and TA_REL static annotations.  AssertHeld actually
asserts that preemption is disabled for the current thread, but only
in debug builds.  A global instance of this class
(preempt_is_disabled) has been introduced and may be used to annotate
functions and methods as requiring that this invariant hold before
being called.

Disabling preemption does not, however, behave exactly like a
traditional mutex from a static analysis perspective.  Preemption
being disabled/enabled is managed using a pair of atomically linked
counters (one for local preempt disable, the other for eager
rescheduling being disabled), and therefore behaves more like a
re-entrant lock, unlike mutexes.  This can lead to a situation where
the compiler might become upset because there exist code paths where
it might be possible to disable preemption twice.  While this would be
a fatal error for a mutex, it is perfectly fine behavior when it comes
to disabling preemption.

So, instead of annotating the existing AutoPreemptDisabler and
AutoEagerRescheduleDisabler with the clang static annotations, this CL
also introduces two new versions of these helpers called
AnnotatedAutoPreemptDisabler and AnnotatedAutoEagerRescheduleDisabler,
which the same way as the original helpers, but also carry the thread
analysis annotations.  The idea is that users can use these version
when interacting with code which requires that they be held for
correctness reasons, and not worry about it otherwise.

Note that this CL simply introduces the token and the helpers.
Subsequent CLs will actually annotate code to indicate where the token
is required as well as ensuring that guards exist in the proper places
to enforce the invariant.

Original-Bug: 100625
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682583
Original-Revision: 58e629b0444e889cf27c97d4963423e82f0e1154
GitOrigin-RevId: 80c6c3ce6a32937b34c539c367300a7e24a3e7c6
Change-Id: I00ecdafed21391121ed611104dee50211b7a286c