[roll] Roll fuchsia [graphics][lib][compute] Optionally check for malloc() failures
This CL is almost entirely a result of applying the following
replacements to the files in `src/graphics/lib/compute`:

* `sed 's/malloc/MALLOC_MACRO/g'`
* `sed 's/calloc/CALLOC_MACRO/g'`

As noted in Bug:97922, for the userland libraries and tests in this
CL, allocation failures are considered terminal and no attempt is made
to gracefully exit.

By default, this CL will make a small improvement and detect an
allocation failure as early as possible but, for now, only in debug
builds.

The macros can be overridden if desired.

Also added a GTEST_ASSERT() to `gbm/test_magma.cc` in this CL as it
was referenced in Bug:97922.

No change in behavior.

Testability: CQ and host benchmarks and tests still run.

Original-Bug:97922

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685564
Original-Revision: 818c55f247f9e43685c9ca46c13b98f0f78f5428
GitOrigin-RevId: 004ed407e47ce5adb00646e6783d151c192dff8a
Change-Id: Ia76c3ac340b0271753eeca9ae76b9be128820b78