[roll] Roll fuchsia [virtio-sound] Add logging
This CL is related to investigation into the root-cause of
fxb/101220 (no sound in Linux VM).

(1) Add logging to the virtio-sound device. Some logging may
arguably be redundant, but is added to make the arrival and
handling of asynchronous events maximally clear to an after-
the-fact syslog reviewer. New logging uses 'throttled_log'
to avoid swamping the log if client is ill-behaved (or
poorly implemented) and is marked by TODO for removal when
the related issue is resolved.

(2) As a preventive measure, when stopping an output stream
the virtio-sound device now follows the pause() call with a
call to discard_all_packets(), waiting for all outstanding
packets to return. This may eliminate a timing window that
hypothetically put the virtio device in a bad state.

Original-Bug: 99637
Original-Bug: 101220

Test: virtualization_sound_tests
Test: virtio_sound_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678464
Original-Revision: bd819362a268fa18efa0236db905a3bbe4132031
GitOrigin-RevId: 0bc31d46cdd8024f7b6ff758422da9eec9962d17
Change-Id: I720b2dc91d299d291104c556a37f0ee2562fcb07