[roll] Roll fuchsia [debugger] Remove support for Component Root
Component root is a concept in CFv1 design and it's no longer relevant
in CFv2 any more. It's also broken after the debug_agent is migrated to
CFv2. This patch removes the support completely. The protocol version
is intentionally not bumped because no one is using this feature and no
incompatibility should be introduced.

This patch also includes some miscellaneous changes to improve logging
and docs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684354
Original-Revision: 946ea97b3baf41bd583acb08c8aa61b915622c29
GitOrigin-RevId: 0b52b5e8629c92ee42676f74124289594d6c3cf1
Change-Id: Ie1d2099847d4e691ea09005276831970945e4e30