[roll] Roll fuchsia [netstack3] Deliver multicast UDP to listeners
Deliver incoming UDP packets with multicast destination addresses to all
matching sockets. This is currently limited to listener sockets, since
it is not possible to create a connected socket that is bound to a local
multicast address. Support for creating connected sockets bound to a
multicast address will be added in a future commit.

Original-Bug: 39479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682943
Original-Revision: d5e7309befbfbbcae3ba1e6a478b0e13456f4187
GitOrigin-RevId: eb42c5c0374f107938e50d55740b04b1d27969a7
Change-Id: I565a07a9a8413cb8bb374a64199f6bc8a819d2e2