[roll] Roll fuchsia [modular][cleanup] Delete basemgr_launcher
basemgr_launcher is used to launch basemgr as a v1 component with a
custom Modular configuration.

This is no longer supported. basemgr should be launched as a v2
component, and any changes to Modular config should be made to
the configuration bundled in the basemgr package.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686573
Original-Revision: 8786b38a57bc58e5a3bd63ba08729a20694bc747
GitOrigin-RevId: 2cacb713b1808257ed580968eff665d58b949987
Change-Id: Iacb73993b679c19ca9aa398ecc81d68e587c05f8