[roll] Roll fuchsia [rfc] Amend RFC-0103 to provide rationale.
RFC-103 defined the situations in which SWD changes will require an RFC.
During recent discussion of an in-work SWD RFC it became clear that the
rationale for requiring an RFC was not clear in some of these,
specifically the meaning and cost of a stepping stone release was not
clear to some readers.

This ammendment adds a short justification for each of the situations in
which a SWD RFC is required.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685372
Original-Revision: 66dde16801c268f995ff6a6572fcaf1c6aded80f
GitOrigin-RevId: 7c1cff81bed2bd96eba2663eb102360a5eaad7ac
Change-Id: I19da9b857f3d3c0a4b2b210217ce30499de8eb6b