[roll] Roll fuchsia [netcfg] Allow forwarding based on device class
This change adds a configuration section to allow forwarding on a device
based on its device class.

A new integration test checks both the positive and negative cases of
interfaces having a configuration installed based on the class. The
previous API implied that extensions to `KnownServiceProvider::Manager`
tests would require additional arguments, but this didn't appear to
scale well in the actual implementation. That's modified to now check
the used config, with the assumption that tests target specific
configurations. This can be improved if netcfg later is able to receive
a configuration as a FIDL message.

Original-Fixed: 94461
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682947
Original-Revision: 180d6b0d3b5be9d697bbf57f0222bbd823fe3971
GitOrigin-RevId: e6ef649cfe974e4421d8db9b95d7ffe025b6e9ab
Change-Id: I237509d640cc7f64b849f09124f8f00c34a4d6af