[roll] Roll fuchsia [scripts/owner] Generate OWNERS files
Extend `suggest_owners.py` to accept an `--generate_missing` argument to
generate OWNERS files for the projects that are missing owners.

Each generated OWNERS file includes the OWNERS files of all the projects
that reference the project.

For now, the OWNERS files are created in the working directory, to be
uploaded to CQ in a follow-up change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686422
Original-Revision: 71bc99e10c193d1e0241848de40ff11c7ea53ad6
GitOrigin-RevId: 4814894905a4ffb5771c68a7520ef33b51271e90
Change-Id: I413bca5fdc41952cb027c78955d8406ec10c236a