[roll] Roll fuchsia [devfs] Make devfs_is_local check more specific
With the check as is, it is impossible to open a device which has an
outgoing directory as it is treated as a local node with the open
request not routed to the driver. This was happening due to
unintentionally inferring something about the device from a flag which
means something else. A new flag was introduced to more specifically
convey the real property the code was looking to identify.

Original-Fixed: 101725
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686614
Original-Revision: ab1255ca8d81d0d6b5c45688ab49b1839ec047ac
GitOrigin-RevId: b790ae9263a81a22bc5b4025a3fb45098dad77e5
Change-Id: I2edc5710d378d07d7ae1b51a875a77f6114011e5