[roll] Roll fuchsia [fidl] Simplify the dyn test suite protocols
Removed three parts of the dyn test suite protocols that don't appear to
be strictly necessary to implement the tests, but add complications
with implementing servers:
- Program point - (the current implementation in Go doesn't even really
  do anything, it echos in a goroutine)
- Observation of StartServerTest - it isn't really clear what you would
  do with this and its existence makes the ordering of initial and
  ending messages less apparent
- Observation of OnPleaseDo - the servers don't even implement this
  correctly today (they don't send anything) and it isn't obvious why
  you want to check that the instruction you just sent has arrived.

Each of these can be added back in if it is discovered that they are
needed, but it seems easier to start with a simpler test protocol and
evolve it as needed rather than guessing ahead of time what will be
needed.

In a similar vein, we may realize in the future that certain features
like MethodPoints for entry/exit are not really needed and be able to
be removed (for instance, we may only need the entry signal).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685268
Original-Revision: f0c2c86a415a10c48b2e3194a2ea8e01585f46bf
GitOrigin-RevId: 34c28483850eaf7546a2718ed193313eff00fa30
Change-Id: If0338b41d4358187c0221e643ee99ab84bec12b7