[roll] Roll fuchsia [assembly] Create 'minimal' product
Factor items out of core.gni to create a "minimum viable" base for a
Fuchsia product, then layer the 'core' product definition on top of
that.

This is not a completed effort, there will be follow-on commits to
move additional packages between these configs as we iterate.

Original-Bug: 100671

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683826
Original-Revision: 73def218dec80a0823decac6ea3213f611af3a64
GitOrigin-RevId: c8c55d9801c146aa72dd77d649806ad637b90319
Change-Id: I884ef69175f2b130bed1fb10a8b6402d28aaca78