[roll] Roll fuchsia [recovery] Do not "re"-load default FontFace.
A request to stop "re"-loading default font was made in fxr/666557.

This CL causes RecoveryAppAssistant's create_view_assistant() to call
load_default_font_face() once.

Copies of the loaded FontFace are passed to ConsoleViewAssistant and
RecoveryViewAssistant's constructors.

Now, instead of loading the Font when rendering a Console or Recovery
view, the "local copies" are used.

For Reviewer: looking for Rust guidance on "It'd be smart to use a
Singleton from Google-friendly crate $XYZ" or "Too much cloning of
loaded font is going on, refactor this to use a ref to a single
object."

If this approach is reasonable, I'll update KeyboardViewAssistant
in a follow-up CL.

Original-Bug: b/233983593
Tested: fx test system_recovery_bin_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685424
Original-Revision: 15a1a2b1e370847a12231ad1dcc40f604b8e5220
GitOrigin-RevId: 2806860d98fd39cbd1a18d1815714104c199436a
Change-Id: I3e494e193b799eae31085da4886e70892bcb1fb2