[roll] Roll fuchsia [brcmfmac][wlan] Return frames on TX fail
In the event that brcmfmac failed to queue up TX frames the frames were
not properly returned to netdevice. This requires the priority queue to
change slightly as it's not safe to handle a moved-from frame if
queueing fails. It also doesn't cover the case where a lower priority
frame queue is evicted, the evicted frame also needs to be returned.

Solve this by providing a pointer to a Frame in the push call where the
evicted or failed frame is placed if needed. This does come at the cost
of an allocation when this occurs. But because we expected users to
provide the TX queue depth to netdevice it is extremely rare that the
queue would overflow so it's better to pay this price only on failure
while reducing the cost as much as possible for the normal use case.
Handle this properly in brcmfmac as well as cases where the entire data
path is down or the SDIO bus is down that would prevent transmitting
frames.

Original-Bug: 101156
Test: Manually ensure no buffer leakage by decreasing TX queue size and
observing queue full events. Included unit test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683822
Original-Revision: 91de702f923855c6cdfb82850f82046e04c2a3f6
GitOrigin-RevId: 0a31fdb2d7cb5fae24854bd7ce56b17fe03ee8fa
Change-Id: Ie0c33a0db292b3d55b71b4ab755562da29cfb7cc