[roll] Roll fuchsia [ffx] Clean up extra build imports
It was pointed out in fxrev.dev/685331 that the components.gni file is
not being used by any of the libraries in the ffx emu plugin, but is
imported all the same. This CL removes the extra imports for code
cleanliness. There is no functional change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687228
Original-Revision: 571bbb586bdd96b81c3d4226df57c8209e84114a
GitOrigin-RevId: cea0b843b7811dd1a2c6bdb4b4b546565828b65f
Change-Id: I8b03dc253f4d4d558bd70ae7564c0434fef03503