[roll] Roll fuchsia [kernel][dev][power] add a quick reset driver for the AS370 SoC
Largely a resurrection of an existing driver deleted in change
21d77ff05881f8af8a5c6e0f87608abd587e91cd

Attempts to use the watchdog driver to reset the system. This is needed
currently due to a lack of a functional PSCI_SYSTEM_RESET hook in the
current firmware. Hopefully will be made obsolete once that is properly
implemented.

Big: 94099

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686763
Original-Revision: fd602683ebf52e08a9f4921028481e1f4e048740
GitOrigin-RevId: cddb4cd0d6d4da6cf3efc4ec3092321e81dd3ad8
Change-Id: I4e487dddd7ba017ed6fb57d700b8940e71fa3505