[roll] Roll fuchsia [staticlints] Check for doc links pointing to Gitiles
Links to docs from source code should link to the doc in fuchsia.dev,
rather than linking to the .md file in Gitiles, primarily because docs
are much nicer to read in fuchsia.dev than in Gitiles.

Any link of the form
"https://fuchsia.googlesource.com/fuchsia/+/<branch/<path>.md" will be
now flagged with a suggestion to replace it with
"https://fuchsia.dev/fuchsia-src/<path>".

This will just be a Tricium warning and will not block people from
submitting links to docs in Gitiles, which is reasonable behavior since
it's not the end of the world if someone does submit such a link, and
there may be legitimate reasons for doing so.

Original-Bug: b/144128643
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686445
Original-Revision: 6a4138549e1532cea6154b0c6ae8e0a08f7257ba
GitOrigin-RevId: 76ccec68d56343baebbf691ffc8215643f3c51b7
Change-Id: Ie34fc0790eee2096d1e06c84009821426a11cce8