[roll] Roll fuchsia [fidl] Convert driver-unit-test to unified FIDL bindings
Currently driver-unit-test uses the C bindings, which are deprecated.
This CL converts driver-unit-test to use the unified C++ bindings,
which continue to be maintained.

The main motivation for this CL was to get rid of usage of
fidl::Builder, which I hope to remove.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687422
Original-Revision: 278e180b6e8e071d78a7e2e7ef5314dbf839e048
GitOrigin-RevId: 9c27b142511f9e440764284bb6ded4d70b700575
Change-Id: I87db7e961f39a9ffbdd11b2375369b60bd2f2608