[roll] Roll fuchsia [fuchsia-fs] rename io_util to fuchsia-fs
fuchsia-fs is a combination of the fuchsia-component/fuchsia-async
naming convention for core fuchsia rust libraries, as well as the
std::fs module for filesystem operations in the standard library. It's
the new location for standard fuchsia filesystem operations, such as
opening files and directories, using fuchsia primitives.

This also includes a redirection for dependencies on io_util, so no
conversions to the new library are made in this cl, to keep it cleaner.

Original-Bug: 36063
Original-Bug: 61251
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687684
Original-Revision: 84d69c2285f0e52bd41d6a4e5ef88ed02adeb4f2
GitOrigin-RevId: f6ea7008b7f88bbae5c503ac83e006d68e865b12
Change-Id: Ifa7510601c30c6cbb440ee29e940f8587dd5534f