[roll] Roll fuchsia [build][tracer] Also detect and ignore temp dirs
Workaround issue where fsatrace fails to report mkdirs.
This messes up the logic used to infer temporary files/dirs.
Since we don't care about dirs (as they have no content),
we opt to ignore them.
If a path is currently an existing directory, it is ignored as before.
If a path *was* a directory that no longer exists, ignore it as well.
File paths are still checked as before.

Original-Bug: 101488

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686462
Original-Revision: 464fa5afbfd3f1feff500bc67033e6318917aefb
GitOrigin-RevId: ac8395f20e91160249726cd3d731ffe075ba19ba
Change-Id: I7043e0c6eeafe056b551dc78d77fc93b40425ae2