[roll] Roll fuchsia [component_manager] Make CollectionServiceProvider more generic
* Rename CollectionServiceProvider to CollectionCapabilityProvider

This reflects that it provides instances of any capability described
by an ExposeDecl and CapabilityName, not just service capabilities.

It also avoids confusion with service instances, which are not
individual capabilities.

* Accept a CapabilityName rather than a target declaration

The declaration is only used to read the capability source name,
which is static, so the caller can pass it in directly and avoid
having to copy/construct the decl.

Original-Bug: 4776
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687104
Original-Revision: 8626b9244cd24cbd23cfd0d148bb95e7ff395072
GitOrigin-RevId: fa52194da44d8b7586f68035477bc88eaac75865
Change-Id: I2b10f4e8683e54a838ab022fbb580c5fd382ec74