[roll] Roll fuchsia [component_manager] Return get_default_provider errors to caller
Ensures that get_default_provider returns an error when it has
a default provider, but can't construct it, rather than logging
and returning None as if it didn't have one.

This makes the routing code fail fast instead of then trying to
find a provider via events, which is unlikely to return anything.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687107
Original-Revision: 3bd9615042d4e894ab15cb09778a7c2b7a043720
GitOrigin-RevId: 6ab1f4fcf03409b6a64bca5dfa3623253598a426
Change-Id: I2f2624d4fe9a1866bbd3c0eb728da70dde587b08