[roll] Roll fuchsia [component_manager] Add namespace directory to RealmQuery
Component Manager will now return a directory representing
the namespace of a component as it exists at resolve time.

The contents of this directory should be equivalent to the
contents in a component's namespace.

This will be used by `ffx component explore`

Test: fx test realm_integration_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687729
Original-Revision: 74d2b72bbf7b5dcb5939705560531254fcf28a17
GitOrigin-RevId: bba9bb9086bdeca33285b863d4b6a18d7534eec7
Change-Id: Ie35161f0a8133d8de3a58c929a2dc58ea7535e03