[roll] Roll fuchsia [ffx][config] Remove magic for tests.
It's much much less confusing to just have tests be clear about how
they're initializing the environment.

This means that if you write a test that uses config and it's run
independently from ffx it will error until you run the test_init()
function, but I think that's better. We could add a wrapper macro
to do this automatically if we really want (and that might be good
because it could also tear it down), but this is a good first step
to that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685323
Original-Revision: 27432bb7b67e9d532cf4a920ef4427d400a17d18
GitOrigin-RevId: d31f90975da8b334296535fc381389c952c3eebb
Change-Id: I6e721a572199da2d1381dddbf0047297d493fb85