[roll] Roll fuchsia [sys] ServeFromStartupInfo should fail if called twice
Change OutgoingDirectory::Serve and
OutgoingDirectory::ServeFromStartupInfo to fail with ZX_ERR_BAD_HANDLE
when the directory request is not valid.

The other option is to follow go/fuchsia-content and crash with an error
message string. However, the existing documentation of OutgoingDirectory
already documented a behavior for this error case, which is to return an
error code. Doing so seemed to mesh better with existing contracts.

Original-Fixed: 92360
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688142
Original-Revision: 8d1118ac6320aa95c178c8329f877832743aedfc
GitOrigin-RevId: d3705410e732785ab13ac7df325aafe66faa50da
Change-Id: Icad5a078b0d27eaf9ea407b4334ebdab52929d9d