[roll] Roll fuchsia [netstack3] Don't require InstantContext for IpStateContext
A later change will only implement `InstantContext` on the non-sync
context but `IpStateContext` will be implemented on the sync context.
`IpStateContext` implements `InstantContext` so it has access to
a type providing an implementation of `Instant` but we can get that
with a type paramter on the `IpStateContext` trait instead.

Original-Bug: 48578, 86168

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687663
Original-Revision: f94509bfc327ebacabacbc2123aa7a13669e6f8b
GitOrigin-RevId: 018b49f8a947c9041314ddfd69a2259c626467fd
Change-Id: I5bf6f99da90b2519266b72d3d387026e19957160