[roll] Roll fuchsia [pi][test] Add a simple profile inheritance stress test.
Add a simple command line program which can be used to stress profile
inheritance in the kernel. The basic idea is to attempt to rapidly
generate many different profile inheritance situations in order to
exercise as many strange edge cases involving profile inheritance in
the kernel as possible.

Note that this is not a stress test.  It does not construct a set of
preconditions, perform a single operation, then rigorously verify a
set of postconditions.  Instead, it relies on ASSERTS and other
runtime debug checks in the kernel to detect violations of invariants
while focusing on producing as many different asynchronous interactions
as possible.  It is meant to be run for long periods of time (days) in
order to flush out rare edge cases, and it has been an important part
of developing the transition from priority inheritance to profile
inheritance.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524420
Original-Revision: 365cf81c7396f91e52a13f12cbdb628c06528425
GitOrigin-RevId: caec2a8183d4a80041acd532f376ec39d51bf8bc
Change-Id: Id5a645336fe6980868b0944721511c332f15ac49