[roll] Roll fuchsia [ktrace] Perform CopyToUser untracked
User copying is not generally allowed to be performed with locks held.
For ktrace lock_ is acting specifically to serialize reads with other
operations, and so makes sense to hold whilst copying. This change
removes tracking for the duration of the copy to prevent the
AssertNoLocksHeld validation from tripping.

Original-Bug: 101783
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687826
Original-Revision: 276b318cd472cbdc4ce2fcb351ba42278d56398c
GitOrigin-RevId: 3c9516d92b3fc733948646ed0fade4061354eff7
Change-Id: I29a99677708e6a1882e0010cc5018a5776e09d99