[roll] Roll fuchsia [starnix] Mount the galaxy and running package inside the galaxy filesystem
This CL ensures that the galaxy package is always mounted at /galaxy in
a running galaxy. Each component that then needs access to its package
because its executable is inside it will be mounted in
/galaxy/pkg/{HASH} where {HASH} is the sha-256 of its /meta/content
file.

For this to be possible a new layered filesystem is created. This
filesystem will delegate to a base one, but have a number of top level
directory that points to other filesystem. This allows to layer the root
of the galaxy image the following way:

LayeredFs:
  base: system image
  galaxy: LayeredFs:
    base: galaxy package
    pkg: tmpfs

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685762
Original-Revision: 2dffdfefdc3058812715be6fc6fca704eacec017
GitOrigin-RevId: 09bbbde1f82491fb8414de8bb5b2790c70a71d50
Change-Id: I42695b246f2880974f5e02dc460e3cee20d4eb7a