[roll] Roll fuchsia [starnix] Refactor DirentSink implementations
This CL moves the responsibility of updating the file position inside
the implementation of DirentSink.

This ensures that the position is kept in sync with the value returned
to user space, and ensures that the file position lock is held uniformly
during the call to readdir across all filesystems implementation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688407
Original-Revision: 54e1c4b1e0856409f3337abd804e94ecbb2bcc4a
GitOrigin-RevId: 9901ddefe2ab56d8b6da7b46dfdb4090bfc5db94
Change-Id: I639997ee5979600767d1b070c3424daa027ee4c0