[roll] Roll fuchsia [fdf][compat] Fix logging on Remove errors
If we go to remove a node controller, but the node controller
has already been removed, the FIDL error will say that the
request has been cancelled.

In this case, we do not need to log an error. In this case, we
are trying to remove our child but it was already in the process
of being removed.

This fixes a flake where a driver calls device_async_remove
while it is shutting down, but its child is also in the process
of shutting down.

TEST: hid-driver-test-dfv2
Multiply: hid-driver-test-dfv2

Original-Fixed: 101688
Original-Fixed: 101311
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686973
Original-Revision: 25771de1a5cb5536df19c88304a88ffdc346eb4f
GitOrigin-RevId: a0681726d994fa78a9e2f7727306af1f63b047a9
Change-Id: I8680d04312e588c9acd97fca57177fab3197e930