[roll] Roll fuchsia [emu] Fix ffx config data for emulator
During recent troubleshooting, we found that the ffx config value to
control the startup timeout for `ffx emu` was incorrectly populated in
the config.json file. The name of the variable is meant to be
emu.start.timeout, with a default value of 60. Instead it was being set
as emu.startup.timeout, with a default of 60.0. This CL sets the correct
default values.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688684
Original-Revision: 295f29d173932e3806f54c760ef42a8466732a2e
GitOrigin-RevId: d002934409b5ee0ec76ac2fc225bc8fdaec0778e
Change-Id: If069836a9ed5b5f35a7e5527ceef1dfa9d131460