[roll] Roll fuchsia [msd-intel-gen] Add callback for connection handle wait
This is used by MsdIntelContext instead of the wait thread,
and makes it easier to support per-engine context queues in
a follow on change.

Release buffer stalls in the connection are updated to process
wait semaphores as required for command buffer submission.

Original-Bug:100943

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680806
Original-Revision: 9442d1dc314714a6673aa33837e017c1ff3aab66
GitOrigin-RevId: 55b8017d42c5a942b533065fcf7006c306f40963
Change-Id: Ib46a73912ac533ba83ca258fdd84fe8955bae383