[roll] Roll fuchsia [libc] Clear start_arg field from pthread struct
Currently, a pointer passed via the start_arg field is accessible during
the entire lifetime of a thread. This can hide an actual leak if the
start_arg was allocated data that a newly created thread never freed.
This CL clears the field right before entering the thread. This pointer
should be available in either the new thread's stack or registers (or
globals).

Original-Bug: 66819
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671534
Original-Revision: fdda85375e8f0273e4c75a228172e783394144cd
GitOrigin-RevId: 10d5a6a4879cb0c3744b2899ca0d945546679584
Change-Id: I8facc97c84d052e118b24e0a10f72f0c10e038e9