[roll] Roll fuchsia [ui][text] Remove text_manager from//src/ui:runtime
Having text_manager be there is an excellent aspirational goal to have.
However, due to unintended interactions with the resolver, this makes
text_manager not be in "base" image in some products. We can't do
that, so to try and fix things, we do the easy stuff first.

The effect of this change should be to move text_manager back to
"base".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688622
Original-Revision: d29a4f0a1d2549a8e90d7983255b8c97e0f5bdb1
GitOrigin-RevId: 66cc8f734b7bc35527ac70900164a5d77e08a72e
Change-Id: Ibb8937398240d9daf5d3d7ebb94db96480f54ed1