[roll] Roll fuchsia [assembly][security] Disable scrutiny verification on boards using custom signing scripts
Disable scrutiny's checks that require parsing the zbi, if the board
uses a custom_signing_script.

A custom signing script is meant to be used when the zbi image must be
post-processed in some manner before it can be used by the target
device.  In those cases, the resultant file is likely not readable by
scrutiny, and so these checks do not work as written.

Original-Bug: 100989

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682824
Original-Revision: a79f5ebbcfe3d3212e54ba2f7b91cc9637011ae5
GitOrigin-RevId: 507254a7ef1f4d2c7f4c1a62d03eb41bb8538e51
Change-Id: I01977e4310082c5b922ca6c391831f9aefd97413