[roll] Roll fuchsia [vaapi] Exposing Y tiling format to client
Allow the VAAPI decoders to expose the underlying Y-tiling format that
Intel platforms process surfaces in. This allows for a zero copy approach
where we will be able to share the output of the DPB directly with our client
and not have to untile the data.

Original-Bug: 99091, 81304
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681765
Original-Revision: 842542eccca809f8986f41c538d59b1bdbdd46d1
GitOrigin-RevId: 2e256878330df6e645a0deb5dbe2893043ccf096
Change-Id: Icb50338dd153a7650c34ade47feadcb3d9b6ebfa