[roll] Roll fuchsia [fidl] Convert ethertap to LLCPP
This is an attempt at constructing a minimal CL that removes C binding
usage from ethertap and moves it over to LLCPP.

In several parts of the CL, it may be possible to move to more ergonomic
usage of LLCPP but this was avoided to more quickly remove usage of the
C bindings and because some quick attempts at simplifying things
resulting it hanging tests that needed further debugging.

The original motivation for creating this CL was to remove usage of
fidl::Buffer.

Original-Bug: 101890
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686972
Original-Revision: a42be1411a3642e52488422bb0dbdf1113cfb9d6
GitOrigin-RevId: 34316dae573cc1a4f3e3e62c921c7f430f4763d1
Change-Id: Iff102f5690b35a764d3a43ac95a4e7af699e7260