[roll] Roll fuchsia [kernel][channel] Restore PreemptDisable during Call and Write.
Go back to disabling preemption during channel write operations, or
the write phase of a channel call operation.  This time, however, drop
the preempt disable if we detect that the list of channel observers is
very long, implying that we might need to keep preemption disabled for
an excessive amount of time.

This allows us to avoid a lock thrash situation which can occur when a
fair thread signals a deadline thread, and the deadline thread gets
assigned to the same CPU as the fair thread.  In this situation, it is
possible for the deadline thread to process the message sent by the
fair thread and attempt to respond to it, blocking on the shared lock
during its write to the channel, all before the fair thread gets a
chance to drop the lock.

In the longer run, it might be better to:

1) Consider ways to break up the locking structure of peered
   dispatchers so that this contention is avoided.  The list of
   observers for the two endpoints of a channel are logically separate
   lists, but they are currently protected with the same lock.  There
   shouldn't really be a need for contention in the case that a server
   thread is responding to a client thread, as the client thread needs
   to wake up observers on the server side of the channel, while the
   server needs to signal observers on the client side of the channel.
2) Consider ways to allow a signaling thread to move the observers
   which need to be signaled out of the queue and onto a temporary
   list before actually signaling them.  This avoids the potential for
   contention, and is arguably better then #1 (since after performing
   its write, a server thread is very likely to immediately wait on
   the channel again, contenting on the lock in the process).
   Unfortunately, it introduces complex life cycle management issues
   for SignalObservers which would affect how waiting works for every
   object in the system and would need to be considered _very_
   carefully.

Original-Bug: 100122

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681805
Original-Revision: 25f64c3e36a3bad9665d0a611c1058817df7ce8f
GitOrigin-RevId: 3c31e7a5b3433bd2a91503024001f58625c8c183
Change-Id: I207d4ea08e1d09c48db4ccaa094ad35fe3cbe5c8