[roll] Roll fuchsia [display][i915] Simplify pipe attachment logic.
The display-pipe mapping never changes once the pipe is attached,
until the display is disconnected; so we can simplify AttachPipe()
by not allowing assigning a display to another pipe. Also the pipe
mode is not set when calling AttachPipe(), so the modeset code in
that function will be never run and we can just remove it; pipe will
be initialized on the first ApplyConfig() where it is involved.

Original-Bug: 102095
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688908
Original-Revision: 667b56b3b30595d01dd5ded5eecf9fbea0ca8054
GitOrigin-RevId: 6c6616bb00fcc5c6f7906f52f2ad19c4c0f19603
Change-Id: Ia5cc6d8021fdaf05b1317cda9ddfcb462413c5d7