[roll] Roll fuchsia [f2fs] Fix a deadlock problem in FileCache
This CL addresses a deadlock problem in FileCache::GetPage(). It could
happen when a thread was waiting for the lock of the n-th Page to be
released after acquiring the tree lock while another thread tried to
acquire the tree lock to get the n+1th Page with the n-th Page locked.
It resolves the deadlock by releasing the tree lock while a thread is
waiting for a locked Page in FileCache::GetPage().

Original-Fixed: 100995

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688969
Original-Revision: 834f86e1ac762a408381b579e9dbc4e93af4ab22
GitOrigin-RevId: 1c25c9aaf4bc91dc9251466548d71a84031f4d47
Change-Id: I8ed78657795d82ebc2b95dd183e17c50c6a1dac0