[roll] Roll fuchsia [ffx] Add pretty_printed aggregated memory profile.
When called with no other arguments, `ffx profile memory` will now
print a human readable digest of the system's memory status. The output
closely matches that of `fx mem`, in the sense that it presents the
same information in the same order, but not necessarily with the same
formatting.

This is intended as a direct port of `fx mem`, and should be more or
less bug-compatible with it.

Original-Bug: 231425139
Test: fx test ffx_profile_memory_lib_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683506
Original-Revision: 5830a2ccba093fc5bf35bd25ce8a2f4f3e42dbcb
GitOrigin-RevId: 458cf13512ea146ce5f5bfb9971ea9bdccc92474
Change-Id: If508b2a8f80acc346b1735c30fadbb5d41e221ac