[roll] Roll fuchsia [fidl] Move dyn suite control channel to natural types
Natural types are recommended for most users, so we might as well use it
for our test harness. A potential risk of this is that there might be a
failure that effects the natural bindings but not the LLCPP bindings,
but they share most functionality other than the type definitions so it
shouldn't be too much of a risk. Also, most base functionality should be
covered elsewhere.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689582
Original-Revision: 6bf36441f8faf7dec0ccc0b4bf8427a709de236d
GitOrigin-RevId: f8de82b1c40b77732cccc1e6517bba9dc3347cb9
Change-Id: I9bdf0de0d3a99cd8f2c0c0f3e9017163f50d5043