[roll] Roll fuchsia Revert "[guest_manager] Remove deprecated dependency"
This reverts commit a40bdbf44ca90782081f5c978f25e8fc12736102.

Reason for revert: This breaks Linux Terminal on workstation. I believe
this is because the previous solution provided a flat directory and the
updated code placed services in a nested svc directory, which breaks
service resolution.

The revert is not pure as the ServiceProviderBridge has been moved. The
only change was to fixup that path and add guest_manager to the
visibility list.

Note: this component will be removed shortly as part of CFv2 migration,
so I don't plan on rolling forward a different fix at this time.

Original-Bug: 102111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689722
Original-Revision: 1e8049f5bd1ef84070666b415b5d6668f6034f03
GitOrigin-RevId: 4b7851704dd66efba7d8c634bf82b8e0f143bbad
Change-Id: I87ca6b795d7c1001b24dc3fbcef56ac95c4f2e9b