[roll] Roll fuchsia [fidl-async-2] Remove server functionality
There does not appear to be any users.

`fidl::BindServer` should be able to cover the majority of use cases.
It's missing one which is putting a hard limit on the number of
concurrent transactions. We could consider adding that feature if
needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687123
Original-Revision: 342b3fe2e31dd861626b1f860aadbd40fb260bde
GitOrigin-RevId: 84649ee36b5d962519e4630b594446b2fdcb8772
Change-Id: I8bb0848b307a75c0b0d46f213e81e345b3cf8b0c