[roll] Roll fuchsia [wayland] Migrate bridge to components v2.
Before this change, linux_runner launched wayland_bridge directly
as a v1 child. If linux_runner were already in v2, we could add
wayland_bridge as a child there, but linux_runner is still v1.

Adds the fuchsia.wayland.Server protocol to the bridging between v1 & v2
and modifies linux_runner to connect to it for each new guest. Instead of
relying on the ScenicWaylandDispatcher to restart the bridge, relies on
Component Manager's restart-on-reconnect behavior.

Deletes the library with ScenicWaylandDispatcher as its behavior is now
provided by the component framework.

Original-Fixed: 100219
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687889
Original-Revision: ea23755855cfab57122d8fb97dac1672a18f6a92
GitOrigin-RevId: 1a9f5cf1bc30df384521f5f082bc64294f2a9240
Change-Id: I8daacbc4c072341503affc0109adaa9b46c6a3e5