[roll] Roll fuchsia [Symbolizer] Clarify format specifiers in spec.
This change makes the meaning of "%i" explicit and removes the implicit
dependency of the format specifiers on C printf/scanf.

The "%x" used as the size argument in mmap is also changed to "%i", to
align with actual usage ("0x" is typically present, but disallowed by
"%x").

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689107
Original-Revision: 3fd0d2359075e00a1c7c5d19251251a0eef62019
GitOrigin-RevId: a78ada418aaf53e4a10db048565667557f9a1403
Change-Id: Ibacf1b398d5a723f926e7de5478dab08d057e063