[roll] Roll fuchsia [netstack3] Return socket state on failed insert
When an insert into BoundSocketMap fails, return the state that was to
be inserted instead of dropping it. This will make it easier to support
multicast registrations, which should be restored with the unbound
socket state after a failed bind.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690142
Original-Revision: b15bd4626f77ebbd99700c4be218b1f8ee1e775f
GitOrigin-RevId: 763040e3785d737f41532b639f19e915b981c75b
Change-Id: I8fd2f54908cea9671f2cda4b67bb66748365183d