[roll] Roll fuchsia [fxfs] Fxfs should check journal_file_offsets entries in superblock.
This change adds a check for object_id that should not be present in
journal_file_offsets. This map contains the set of offsets that are
yet to be flushed, keyed by the pinning object_id.

Entries in this map that shouldn't be there may make it impossible to
free journal space.

Note that this originally had an associated test but I've dropped that
as Journal is not accessible via FxFilesystem and the super block is not
directly mutable. I didn't want to introduce leaky abstractions for the
sake of this test. I also tried inducing journal reclaim and device
flush but because a device flush tries to first flush the ObjectStore,
I could only hope to probabilistically land an entry in this map by
issuing writes at the same time a filesystem sync and this would
inevitably be flaky.

Cost/benfit seems negative to me, thus no test of this check here. :(

Original-Bug: 101660
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687382
Original-Revision: 555f7ec4d8a7c24bf6a20d077451a17b2d5dc15e
GitOrigin-RevId: 79f22b408fc73de982568dc624977371a0af39bd
Change-Id: Ib9a04fb5e5afce2988fc83416610de59a1439749