[roll] Roll fuchsia [fint] Never consider installer_test affected
It depends on a system image so it is practically always considered
affected by changes. It doesn't matter that much because it only runs on
CI-only "slow" builders, but it's confusing for it to show up in fint's
affected tests list nonetheless.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690105
Original-Revision: 36b35d2b965cafb420914108cd8db850efe95ace
GitOrigin-RevId: 2f66c8d2f6eebdb34960fe3daa79a2965d756494
Change-Id: I561ddb07e04f40e4903b07f711bc307552a438c5