[roll] Roll fuchsia [fint] Fix bug in affected BUILD.gn file analysis
When checking whether a test was affected based on whether its BUILD.gn
file was changed, we were previously comparing the _absolute_ path of
the BUILD.gn file ith the _relative_ path of the test's GN label, so we
were never considering any tests to be affected by their GN file
changes.

Now we always use relative paths to compare test labels to changed
files.

Also simplify the `removeDuplicates()` logic as long as I'm here. It was
optimized to avoid unnecessary allocations, but that made it
significantly harder to read and made me suspect that it was concealing
the bug that was causing this issue. It will never be used to sort more
than a few thousand elements at a time, so the optimization was not
worth the loss of readability.

Original-Bug: 101789
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690149
Original-Revision: 5dd3360ebc12778402016362bd0b2405c1b586da
GitOrigin-RevId: 6889d554ddb5720ab4eafcbe6dc0e9ba907854c6
Change-Id: I5b2cb538cde14675f54456d3d56d5405da15a6c9