[roll] Roll fuchsia [upgrade_test] Double cycle timeout
The sherlock-release-ota-min-version_system_tests_upgrade test has been
hitting this internal timeout consistently since this morning, but it
just seems to be running slowly as opposed to encountering a hang.

Bump the timeout for now to unblock CI/CQ until we can get to the bottom
of the failures.

Original-Bug: 99533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690492
Original-Revision: 13f6898b4b2229198553cc7e49b16775b85eb4fd
GitOrigin-RevId: 118f14476f8a4a2af9908ba3b130498005469f37
Change-Id: Ib8418472cda99780457cce2b4bfcb81be96e7b6c