[roll] Roll fuchsia [kernel] More precisely conditialize '64-bit' printf test cases
We were incorrectly relying on the presence of the _LP64 preprocessor
variable to establish whether certain integral types are 64-bit (e.g.,
size_t). In actuality, this variable is only defined when `long int`
and pointers are 64-bit _and_ `int` is 32-bit - and so is a poor test of
whether, say, size_t is 64-bit. As a concrete counterexample, the x86
EFI target has `long int` as 32-bit, but `size_t` as 64-bit - so in
particular, this change allows the printf tests to pass in that
environment.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690162
Original-Revision: 88dcf93e5aacd6c52f8fcc3efcc81751d41fd070
GitOrigin-RevId: 00e15c4b651078502d1bdcf586c8a124536cf507
Change-Id: I60a909891135be53a4719b9c8259a9d7588fb4af