[roll] Roll fuchsia [vaapi] Share tiled DPB directly with clients
Instead of copying DPB data to VMOs have the DPB be backed directly by
VMOs in the tiled case. This way we can share directly the VMOs with the
client and not have to copy. To accomplish this and maintain
compatibility with the linear format, a SurfaceBufferManager interface
was introduced. Each version of this interface must handle the
relationship between packets, buffers, DPB and VMOs.

Original-Bug: 99091

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682363
Original-Revision: 87f534efd370f05fef6af4aa917d703c9092c028
GitOrigin-RevId: 73420135e0917e1be88064d03f77deb4c8bf66db
Change-Id: Ie08a85ec463117802cb29751ec3949ac9fbf6596