[roll] Roll fuchsia [docs][style] Revert change to line width.
The line width of 80 was changed to 100 in
https://fuchsia-review.googlesource.com/c/fuchsia/+/498703
which misunderstood the difference as an oversight.

I'd make a revert for that here
https://fuchsia-review.googlesource.com/c/fuchsia/+/682525
but the file has changed to where a simple revert no longer applies.

So this CL is a direct change, along with more explanation (to help
keep this from changing again). But in some sense, this is still a
kind of 'revert' of 498703.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690522
Original-Revision: 5dbb7ba01ea7e0425d52c6016b7caa5d7015de5b
GitOrigin-RevId: 0eeed4b8c42f736a79f1efd96ef2a147304554e4
Change-Id: Ib9594704dfcfa642c8237fd557e6a39f7cb90b85