[roll] Roll fuchsia [ffx][config] Collapse and simplify Config.
This is a rather drastic refactor of the implementation of
ffx_config::FileBacked, which had a complicated layered architecture
that made it difficult to reason about. There are likely reasons
it was made this way, but I believe they've largely been forgotten,
and even if we re-discover them I think we'll want to do it
differently.

Among this there are a very few small behaviour tweaks, but I have
tried to refrain from making any substantial ones. Some things that
were passed in as references to options are now passed as options
of references, many impl-level functions were moved to globals,
the nested_* functions were moved to mapping, and the tests
were merged into one set of tests.

Also the logic for validating whether a write-operation was allowed
was brought out to its own function. Collapsing this logic being
done at every layer means that even at the former Priority layer,
you can't set Default level values, which I think is correct,
but the tests had to be changed to accommodate that. As far as
I know this is the only behaviour change in the tests, and it
was on a part of the code that was not previously visible to
callers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689804
Original-Revision: 34cdd988d771efa1ffed1ea59d78fd83080a7865
GitOrigin-RevId: dc80547768346d94849a589b682efd088b61d3d5
Change-Id: I2d9791675bf2312eae7f86d9be3b846b963f6cb6