[roll] Roll fuchsia [mdns] remove additional empty string from generated TXT records
This CL fixes a bug in the DNS message writing code that added an
empty text string to the list of text strings in a TXT resource
record. An empty text string is still added when there are no other
strings in the resource.

Original-Fixed: 102543

Test: new unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690412
Original-Revision: 91148002e97853cf18aa8798579b274919f72858
GitOrigin-RevId: 9988429feaaa4789957ff8e54768ee06321e68f1
Change-Id: I85351eee5adaa67a3a995e7240a34bca3557f843