[roll] Roll fuchsia [bt][gatt] Clean up MTU exchange result handling
Instead of returning a tuple of (att::Result<>, mtu) from gatt::Client::
ExchangeMTU's MTUCallback, return an att::Result<mtu>. This represents
the result of the procedure more accurately - if the MTU exchange fails,
no MTU is negotiated.

The one edge case is if the MTU exchange fails because the peer does not
support the procedure. This procedure is optional per v5.3 Vol. 3 Part F
Table 4.1, and in this case we should continue initialization with the
understanding that the MTU is the default, minimum LE MTU. This is the
only behavior change from the prior code, the rest of the change is
effectively a refactor.

Also, change some EXPECT_EQ->ASSERT_EQ in RemoteServiceManagerTests so
that failed expectations do not cause invalid array accesses and process
termination.

Original-Bug: 36375
Test: `fx test bt-host-gatt-tests`, refactored gatt/client_unittests.cc
      and added
      RemoteServiceManagerTest.InitializeMtuExchangeNotSupportedSucceeds
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689574
Original-Revision: efd151c3a1282a48157b125354ac7db744e8ef0f
GitOrigin-RevId: a9ef5dd27769ef2ad087011dfeab9010d95b524b
Change-Id: I86ae6b4fa157fc2b069b16c8bf459170951aa120