[roll] Roll fuchsia [wlan][policy] Explicitly select authentication method.
This change constructs an `Authentication` FIDL message in the Policy
client state machine when connecting to networks. To accomplish this,
the configured security protocols of scanned networks is included in
`ConnectionCandidate` and used to select an authentication method.

Some information is only present in a `ConnectionCandidate` if a scan
has been performed. Since this change introduces more of this kind of
data, these fields have been refactored into `ScannedCandidate` and a
single optional field of this type.

The `Authentication` FIDL message is not yet used and this change
temporarily discards everything but the bare credential.
`Authentication` will be used directly following changes to the SME FIDL
APIs at which point this temporary code will be removed.

Original-Bug: 72051
Original-Bug: 72906
Testing: Unit tests and manual testing.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672744
Original-Revision: c6c17c18d0368442a8fc73e4e9aeb9277df9d5fd
GitOrigin-RevId: 4ef74bb401eecd4f950997058e8cd63ac1261bc7
Change-Id: I455ca974f9345c2bb15023b6ab4c00671b4d1a6c