[roll] Roll fuchsia [docs] Fix header names for syscalls.
The NAME header is for a section containing a summary
of the syscall. It does not contain a name.

The SYNOPSIS header is for a section containing the
C function signature of the syscall. It does not
contain a synopsis.

Changed NAME -> SUMMARY.
Changed SYNOPSIS -> DECLARATION.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689004
Original-Revision: b715804372cb442b44cec4ff93bfbb09b664e547
GitOrigin-RevId: 7ea4e5df885292a1577f11c88dbf3c7d2f033bc8
Change-Id: I655a9591ae15ffa605d5c0385016f23396d48e0d