[roll] Roll fuchsia [fxfs] Roll the journal mutations key
The Chacha20 variant we are using only allows a 32-bit block
offset (where blocks are 16 bytes) which isn't so big that we need to
not worry about wrapping.

To address this, this CL will roll the mutations key after it has been
used for 128 MiB.

In addition to this, this CL tidies up some other encryption related
issues:

 1. The encryption keys are now a separate key/value pair which avoids
    having to update it every time the allocated size or reference count
    changes.

 2. Moved the key ID out of WrappedKey and UnwrappedKey since that is
    specific to keys used for extents.  WrappedKeys and UnwrappedKeys
    now store tuples.  This means that we can use WrappedKey and
    UnwrappedKey consistently.

 3. EncryptedMutations can be bigger than the default maximum
    serialization size allowed.  It used to fail if we exceeded that
    limit.  This CL fixes that.

As this is a breaking change, I've removed support for all older
versions.

Original-Fixed: 93354
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685705
Original-Revision: cead6fb78ce658308d0b99962a4873cbbff3d54e
GitOrigin-RevId: e9e24024a3c5dece60671caaeab491ab2e905e7a
Change-Id: I27d2c1b69c5e13e52f5b238771801e4c114e8e1d