[roll] Roll fuchsia [net-cli] Make net route --metric optional
The fuchsia.net.stack.ForwardingTable FIDL definition documents a
`metric` of 0 to mean that the metric is unspecified:
https://cs.opensource.google/fuchsia/fuchsia/+/main:sdk/fidl/fuchsia.net.stack/stack.fidl;l=28;drc=2707f1db129a8c6fa20b5a5aa282fde9dcd400dd

We should expose this fact in `net-cli` for the convenience of
network-conformance testing scripts. `net-cli` used to expose access
to an equivalent `fwd` command which did allow for omitting the metric,
but the `fwd` command was removed in
https://fuchsia-review.googlesource.com/q/I71326cd4c894f88e2d4aae524433e69c853c532a.
I overlooked the fact that `metric` was made non-optional when migrating
testing scripts to use `route` instead of `fwd`.

Test: net-cli-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690822
Original-Revision: 0b9e411517e95eed51ecd185775a3908982eca4d
GitOrigin-RevId: 251b91ca8072c4f3c8eedd0f8e5995246b276654
Change-Id: I79df5178d056fe5d383fb888999427d60fdb34e3