[roll] Roll fuchsia [fidlc] Add compiler support for new-types
This CL adds support for the new-type syntax in the compiler. This
patch is responsible for adding the code that will handle new-type
declarations and a flag to gate this feature. The code for emitting
the correct JSON for the IR will be provided in a dependent CL.
Similarly, the language specific binding generation code will be
provided in separate CLs.

Original-Bug: 7807
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681402
Original-Revision: 16d3790b08445af0521e34eb1590ea905a36c6c0
GitOrigin-RevId: 6c297f995654ae30c1dd2cb29f2c77adcf3ddf09
Change-Id: Ifd5f331d4146eb4432d587c0d955725f0d134df9