[roll] Roll fuchsia [fx][rbe] fx set --rust-rbe, use rust_rbe.enable
fx set now provides `--rust-rbe` as a renamed version of `--rbe`
(to be phased out).  `--rbe` continues to work (non-breaking).

Use the new variable `rust_rbe_enable`
as a replacement for the old `enable_rbe`, which will be removed
in the future.  We print a warning to update for now.

This will avoid confusion with a future `--cxx-rbe` flag.

Original-Bug: 101917

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688147
Original-Revision: 69c35ea94d919a52c7561d79034c621f5575201e
GitOrigin-RevId: d311f5b47660588981277eb5287ab6f2b43449ce
Change-Id: Iae61d51543c3930a49673ba0066c78f98f2b978e