[roll] Roll fuchsia [msd-arm-mali] Put GPU in nonprotected mode before perf counter test
The perf counter test expects the GPU to be in unprotected mode so it
can write out the performance counters. Some earlier tests may leave it
in protected mode, so we need to send a non-protected atom to the GPU to
force it to always switch to non-protected mode.

Original-Fixed: 102538
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691122
Original-Revision: dd5cbf5f479eb44928cb64364191383131ab5029
GitOrigin-RevId: afb38a970a070cf62f6f617245e56df8215bde29
Change-Id: I4d5d9c309b01542f232cf91d4dd6466a829dccdc