[roll] Roll fuchsia [identity] Delete TokenManagerFactory.
TokenManagerFactory was intended as a way for products to integrate
with Oauth and OIDC providers without using the full account system. We
don't have any products that currently need this capability nor do we
know of any in the future.

This 2018 implementation has not been maintained and would likely need
an overhaul before it could be used and the reality is that products
would be as likely to bring their own implementation as to use one
provided by the platform. Deleting it now saves the ongoing cost of
maintaining it and in particular the need to migrate it to CFv2.

Original-Bug: 101940,101941
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690673
Original-Revision: 2bf57e3cac76fa881bbb18fe9270ce3128b71648
GitOrigin-RevId: b04007e6bb91b2e7b071374609101e6edb77d9c6
Change-Id: I2ca014d902ef138e209c901e00410e364223d835