[roll] Roll infra/recipes [testsharder] Allow Multiply count immediately after colon
Previously the "Multiply" syntax "test_name:123" (where 123 is the run
count) was not allowed, because when space after the colon is not
required, it's hard to write a regex that disambiguates the colon from a
colon that might appear in the test name.

I fixed this by making the test name match non-greedy.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/691188
Original-Revision: 44ec809aaf672f9ebe5b7e534b70f7b5b261864a
GitOrigin-RevId: 66b278f60b8cddd44cfa1e56c2f9d85b0d6399e7
Change-Id: Ib50b6601365181e430c61d2a1c22defcf380eb3c