[roll] Roll fuchsia [boot-shim][depthcharge] Specially configure the UART
This is effectively a reland of fxrev.dev/683850 for everything not
loaded by depthcharge. It introduces a hack for the depthcharge case,
which will soon be removed in favour of a proper CMDLINE-bases fix
described in the associated bug.

Original-Bug: 101395, 102726
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688826
Original-Revision: 0a4ebbb34d07672efa017e399069a03eb6d3e974
GitOrigin-RevId: f275251a47dc759fbbeb3421a1bfbc8856f1b3b7
Change-Id: I7b696f7d00a68d12e68fe107dbafc78a975757fa