[roll] Roll fuchsia [ot-radio] Changes to fix ot-radio TX delay
There is an issue in the current handling of interrupt which results in
hold-off of ~40msec in packet to be transmitted. This manifests as a
test failure in bug 231208021, but affects most SPI transactions.
It is found that TX was held off due to RX taking long time because of
the delays in each interrupt. Modified delays to make faster exit from
interrupt handling.

Currently, removing delays for first 5 do-while attempts in receivng
transaction. This takes care of the common case where 2 transactions are
needed for receiving long message - first transaction confirms the size
of inbound packet, and second transaction is done with the size. This
was taking ~10msec at ToT because of the delay in do-while loop.

Also added an early exit for spurious interrupt event. Since the inner
loop can take care of multiple interrupt events (while handling the
interrupt the pin gets deasserted and reasserted if packet is still
present on the other end of SPI). This results in an interrupt event
getting queued up.

Original-Bug: b/231208021

Tests: 1) ot-radio driver test.
       2) Thread Harness joiner and Leader test cases.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690346
Original-Revision: 84ac3e62ee109867a99148720cdf85e8c4b2b780
GitOrigin-RevId: 3fb335b523cebd952d55a5d5f94d52f7de861d12
Change-Id: Idba05573141c597c640d41550b71c71d1ec7cd13