[roll] Roll fuchsia [minfs] introduce minfs component
This introduces the minfs component. Similar to the blobfs component, it
starts a fuchsia.fs.startup.Startup protocol when the component is
started, which is used to further parameterize it with the block device,
options, and operation to perform. The lifecycle protocol is also
implemented.

This adds two new sets of tests, one which tests the ComponentRunner
implementation directly, and the other which actually launched the
component with the cml as a dynamic child and calls the startup protocol
on that.

Original-Bug: 101655
Test: fx test minfs-component-runner-test minfs-component-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686914
Original-Revision: 611714433ed0182444cd3e85127f0eca24f29686
GitOrigin-RevId: 065c2fa0c61c75e02ba297535a5ffc8bbdacfe31
Change-Id: Ibd7004136372d500a34275b1a091e5089b4ef073