[roll] Roll fuchsia [scripts/owner] Upload OWNERS files for review
Extend `suggest_owners.py --generate_missing` to commit the generated
OWNERS files and upload the change for review. Add as reviewers all the
newly added owners.

Add a `--dry_run` argument to print the content of the generated OWNERS
file and the git commands without creating the file or uploading the
change for review.

Ask for confirmation before uploading for review if it's not a dry run.

Drive-by: write to cvs file only if `--cvs` argument exists.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691402
Original-Revision: 8fa1aeac0076ac97521050330a84a0866ea9738a
GitOrigin-RevId: 7317954f3fbf938eb001e8b49d8fe89f13564c07
Change-Id: I64c5b7fc8dd1a02bd2988a8ca9ad4b4df0e40866