[roll] Roll fuchsia [debugger] Make stack checking more paranoid.
I got an assertion during normal use that there were no stack frames
when we expected there to be some. I was not able to reproduce the
problem, but this adds stack frames to an extra case we were not before.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679456
Original-Revision: 1a0331c05ac43001ccdb05ad38121c1594bf01b4
GitOrigin-RevId: 21beedab15aed58eaccddfac0462e666912027ca
Change-Id: I345519e53b5e4b5d9ac2e5f174d446b84f023678