[roll] Roll fuchsia [vulkan][compact-image] Fix decommit on trim
If there are multiple imports of a buffer, each of them most be unpinned
before the decommit is legal. This happened to work before because
within a single driver multiple imports were being deduplicated, so an
unpin of one unpinned all of them. That doesn't happen anymore.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691748
Original-Revision: 293dde0954bbfc4af945feecc227211da753c15e
GitOrigin-RevId: 1094ec434cad852a6c8e9fb9848f666275e86a6c
Change-Id: I55e338f162629c64bbd3d4f6c1410f05dc28cd1e