[roll] Roll fuchsia [network/tests] Introduce multi-nic test environment
This introduces a multi-homed network environment that tests may run
on to validate socket operations when a stack has multiple
interfaces/routes to a destination.

A client environment is created with 2 interfaces (on the same broadcast
domain) and a server which will simply respond to any received UDP
packets with the same data but with "Response: " prepended to it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691306
Original-Revision: f6be0256e434a14846b33bc30c4f38fa1c843cac
GitOrigin-RevId: 0e782616efc3b77ea0ee9dc7d67c518c1fb00165
Change-Id: Ia3a3c341b2c59f87c2431f07f4f48a729e135252