[roll] Roll fuchsia [docs][hub] Remove 'deleting' dir documentation
The Hub's `deleting` directory was removed in Change-Id
I58da854b73981838799004f1c89002e435619c0f.

This CL removes references to the `deleting` directory from Hub
documentation. There is no longer a "staging" directory for a node that
is to-be-deleted in Hub.

The reason for this CL is update documentation from changes made to Hub.

Original-Bug: 93481
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691843
Original-Revision: 880888448f2ca802903227268949785897ded208
GitOrigin-RevId: 067541bd2a19745d3536b077d82445d9deae5bf7
Change-Id: Id4de63a7276453c4213e74a07f9ee8c100b86326