[roll] Roll fuchsia [socket] Add functions for reacquiring signals
There's a Rust socket optimization where READ/WRITE signals stay
set until a read or write fails.

Some applications require that polling for socket readable or socket
writable actually means that the socket is readable or writable before
trying to act on it, so an API is needed that clears and reacquires
the signal lets us poll and know that the signal is "fresh".

Example use case for virtio-vsock:
1) Guest tracks socket buffer and never sends more data than the socket
   can handle, so a write never fails and the signal is never cleared
2) The device wants to wait on socket writable to inform the guest that
   there is room with a credit update
3) Polling always returns Ready since (per 1) the signal is never
   cleared
4) We tight loop spam the Guest with credit updates telling it that
   there's no credit

The solution is to have the ability to clear the signal before polling,
so we know that the polled signal is current. This won't be used
generally as most applications can just let the write fail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691004
Original-Revision: 548c0d029108af115fb5cf8165798466a78440f5
GitOrigin-RevId: cb127f887ac892f552fb1b9af89ff4ad1b0ce8f7
Change-Id: I80d2774c25a3b4ad0c59fb0bb94c41256f3742f0