[roll] Roll fuchsia [testsharder] Add max shards per env test case
It's important that testsharder try not to exceed `-max-shards-per-env`,
so add a test case that covers the path where it needs to recalculate
the target shard duration to keep the shard count low.

`-max-shards-per-env` is technically not a hard maximum and we will
slightly exceed it for environments that are split up by affectedness
and hermeticity, which is why there are three non-skipped shards in the
resulting golden file even though `-max-shards-per-env` is 2.

Also add a `testListEntry()` helper function to make it more concise to
mark tests as hermetic in the test data.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691271
Original-Revision: a0b56aece64229adaf2999d0a757505a9c367f51
GitOrigin-RevId: 3c553cb7e324730069d9b97509a276e6e3497820
Change-Id: Iedb1fec2ad79d8a4b248c8732cd40b75982e93c0