[roll] Roll fuchsia [netstack] Add trace points at netstack boundaries
Introduce trace points at the boundary between the netstack and client
applications (through the `fuchsia.posix.socket` API), the boundary
between the netstack and gVisor, and the boundary between the netstack
and drivers (in the ethernet and netdevice clients).

For now, only async events are emitted. NB: it's necessary to use async
events (as opposed to regular trace durations) because the Go runtime is
free to move goroutines between the backing OS threads, and asynchronous
tracing is necessary to correlate the start/end of an event that spans
multiple threads.

Also, no flow tracing is used to correlate the flow of a particular
packet through the system; this might be added in the future.

Test: Follow instructions in src/connectivity/network/TRACING.md to
      enable tracing in the network realm, and run `ffx trace start`.
      Verify that the resulting trace includes events in the `net`
      category.
Original-Fixed: 99538

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688426
Original-Revision: 40e3ef4c0e46a0167266afc15e5a408e9e7d2d54
GitOrigin-RevId: cc39bb91cd95584d56ffa05563372d3ebc232540
Change-Id: I500e701bf0c7dfd0c29c1d1cec714a7946dc7d6b