[roll] Roll fuchsia [slimfast] Memory budget boot test
Introduce a host-target test that boots Fuchsia to a session in a
RAM-limited VM.

In the course of this work, discovered that fshost doesn't seem to be
able to mount a virtual disk pointed at `storage-full` image (a raw FVM
block image) because it runs out of space. Turns out our dev tools
already handle this by creating a temporary upsized image when the tool
runs:
https://cs.opensource.google/fuchsia/fuchsia/+/main:tools/devshell/qemu;l=141-147

This patch also adds a mechanism for tests to do the same.

Original-Bug: 100672

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683825
Original-Revision: 36d05ab2497652e32b85538161b1d69e464276fb
GitOrigin-RevId: 09970fffc96cbcd8f7f29b0b57aeb96213ecbdf8
Change-Id: Ia9c12c972ef420acb365164c246db44cf07edb5e