[roll] Roll fuchsia [ot-radio] Fix inadvertent interrupt signalling
Found that InvokeInterruptHandler was called in normal operation. This
was supposed to be called only when we stopped handling interrupts
because allowance went to 0. On a subsequent increment of allowance,
we wanted to trigger Interrupt Handler. But it was getting called
inadvertently when we set the flag to true in the inner do-while loop,
and in next iteration of do-while loop, another thread increases
allowances, find the flag to be true and Invokes interrupt handler.

Also made the boolean atomic since it's shared between the two threads,
and writing to it only if there is a change.

Original-Bug:  b/231208021

Tests: 1) ot-radio driver tests
       2) Thread Harness joiner and Leader test cases.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691525
Original-Revision: afda3876c099f0ba679dce4552ed820f8477d653
GitOrigin-RevId: 728b5dfac3e7760ae451e6ca46de4ba2c40a25a1
Change-Id: I0696ac71e7aadf3862f1daa64d9124344cafd7ae