[roll] Roll fuchsia [tefmocheck] Allow for truncated except block.
The end string of an except block may appear after the last test
finishes and accidentally get truncated from the logs since we don't
have a way to know to wait for it. In that case, we should assume that
if we see the start string of an except block in the logs but no end
string after, it was probably truncated and so everything from the start
string to the end of the log should be included in the except block.
Except block strings should be chosen carefully so that you wouldn't
normally see them in the logs except as part of an except block.

Original-Bug: 102747
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692168
Original-Revision: 25b80535186e518567664fc89bd7014f523adf23
GitOrigin-RevId: 25683c603dd743865b1c02d1d23cf36a795f9259
Change-Id: I0084b0660c39a2cbaf811165e5bc5f2fa144f43c