[roll] Roll fuchsia [netstack] Benchmark socket API with fake netstack
Add a variant of the socket benchmarks that runs against a "fake
netstack", which does the minimum possible amount of work to implement
the subset of the `fuchsia.posix.socket/Provider` protocol needed for
the benchmarks.

By comparing with the same benchmarks for the production netstack, this
will give us a signal about how much of the latency in e.g. UDP is due
to work done in the core netstack vs. due to the overhead of the API
(IPC, kernel roundtrips, FIDL de/serialization, etc.).

The fake netstack currently implements support for UDP and ICMP echo
sockets. Support for TCP and potentially "Fast UDP" will be implemented
in future changes.

Original-Fixed: 98501
Test: fx test socket-benchmarks-with-fake-netstack
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670947
Original-Revision: a95660dab37d13272feed8b1dbfd8d91c545c43b
GitOrigin-RevId: 6515400e8c0d978f388b098c764d63e532bd4fea
Change-Id: Id27f37ab98eb615febf503f4a56534e152d2c460