[roll] Roll fuchsia [blobfs] Change blob loader so that it no longer attaches VMOs
Calling AttachVmo currently has the potential to trigger a deadlock with
the driver stack if something else is calling GetTopologicalPath on the
same driver host that contains Blobfs's underlying block device.

To avoid this, change BlobLoader so that it only attaches at mount time.

An unfortunate side effect of this is that it will involve copying
merkle tree data. However, this is more than likely offset by the fact
that we don't have to wait for an AttachVmo round-trip when loading
blobs.

This change may have a positive impact on memory since memory-resident
merkle trees won't have to be padded to the system page size.

Original-Fixed: 102743
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692482
Original-Revision: 11108b067c5cf5ea77fd1bcb0da90287de118c86
GitOrigin-RevId: 77d40672c3d9a82021c81b42dac53d1b7c3c1dc3
Change-Id: I19291a6124853c799cca5e502f3fd518121538d8