[roll] Roll fuchsia [intel-i915] Rename display_count -> display_config_count; use std::span.
The name "display_count" used in intel-i915.cc is misleading: usually
it stands for the count of display configs passed in by
ApplyConfiguration instead of the actual number of active displays /
pipes, which could cause error when we calculate the DBUF size
(will address in a subsequent CL).

This CL just renames all the "display_count" to "display_config_count"
to reflect this fact, and replaced a lot of (raw pointer, count) usages
with std::span, which is more clear in semantics.

Original-Bug: 95863
Test: Variable renaming + minor refactoring. No-op.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691530
Original-Revision: 8f85dd84ed697a64c595aeeef0aa68d33d22ecb5
GitOrigin-RevId: 290354dde513c467fc01f8a7ae8b6e8c582de837
Change-Id: I4539a6fa16367481e0771dfaf028e44f539cb950