[roll] Roll fuchsia [ffx-emu]Adding engine_type property to runtime config
This exposes the engine_type as part of the runtime config object
so that it can be used when special-casing flags based on the engine.

A concrete example of this is touchpad input is currently only supported
on aemu, but the goal is to add support later so this should not be
hard-coded into the launcher.

Original-Bug: 99762

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681563
Original-Revision: 7f5d3e9e18e3951c10ddc0f05ca9a0f0b1303014
GitOrigin-RevId: 3eeb0a42f501e4564d024d7646d9f3d9995dc871
Change-Id: Ib05800e4e071be9d5c13d3fd989baf56becb5a82