[roll] Roll fuchsia [kernel] Change OOM timeout defaults
Increase OOM timeout default to 50 seconds.

For the bringup product set the timeout to 8 seconds. Bringup is
a small product and we'd expect reboots to be quick, if it takes
longer than 8 seconds, something is very likely wrong.

For the core product use the default.

Userspace now shuts down in an orderly fashion and is expected to
restart the device promptly so the kernel timeout functions as a
backstop.

Empirically all product configurations almost always reboot in
less than 10 seconds so we believe if the system actually takes
greater than 50 seconds something is truly wrong in userspace.

Original-Fixed: 91744
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692170
Original-Revision: 529568a4f7562ad7f5e66fad416fa3b77a02545c
GitOrigin-RevId: ad01d09d73d2d4ff5f6ab9489d57707df0543199
Change-Id: I642416ce6d78bd2837de8d7606dfb5d649531cf5