[roll] Roll fuchsia [process_builder] Fix reservation vmar underflow
When creating a reservation vmar the size is calculated as:

((info.base + info.len) / 2) - info.base

This can underflow if the lower half of the address space does not
intersect the main vmar. This change updates the logic to always make
the reservation vmar half the length of the main vmar.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673425
Original-Revision: d0551c22ac46401415e15518354ec16e48bfa3f7
GitOrigin-RevId: 703f5e237581682c662b4c25915d4d7c70e549ef
Change-Id: If5b70fdf02f9fbc145a42eefe5b590c9f6b99ba4