[roll] Roll fuchsia [fshost] launch minfs as a component
Make fshost launch minfs as a static child component. This mostly
piggybacks on the fxfs fshost changes, so it's mostly a cml change, and
changing filesystem-mounter to call the component launching logic.

This also adds a change to the minfs component runner to skip some steps
on shutdown if we don't have a valid pointer to a minfs, so we don't try
to call things on a minfs instance that isn't configured yet.

Original-Fixed: 101655
Original-Bug: 91577
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686949
Original-Revision: cf56375f923deef12ac51ac69df26dafa7d2e770
GitOrigin-RevId: d6e127392733113b416b629adcd9f2e89603ccf5
Change-Id: Ie8d234a73935cf5d745602e9988ab041352f662d