[roll] Roll fuchsia [amlogic][i2c] Add a metadata field to specify the SCL low delay
The bus frequency and duty cycle are determined by the quarter clock and
SCL low delay values. Add a metadata field to specify the SCL low delay
to use, and populate it with the values to correct the I2C frequencies
on Nelson.

Original-Bug: 102864
Test: Exercised all I2C buses on Nelson
Test: device-enumeration-test passed on Astro and Sherlock
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691475
Original-Revision: d60bdceac25cdfc43b54c280ee19d40865a42341
GitOrigin-RevId: 9bd076ee03b2a5a36e66497d144bf5777f94e4b8
Change-Id: Ie4e2d5158557c8e61a54ed2885b28a06269d6971