[roll] Roll fuchsia [fidl] Remove fidl::Builder
This CL removes fidl::Builder which generally shouldn't be used by
clients and has limited usages. fidl::Buffer is functionality that
makes it possible to avoid certain allocations, but if this
functionality is needed LLCPP should really be used instead.

The original motivation for removing this was because some usages of it
made it harder to move HLCPP away from coding tables and due to my
surprise after seeing the fidl::Buffer name and not knowing what it was.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686580
Original-Revision: 540f16781c2b92dfaba2a691eb41e8b707731c63
GitOrigin-RevId: 53369daa3425a6d24ebae4cf44f302d555a12520
Change-Id: I85a34ebe9544051b05b5d1515887f5562e5dd7c7