[roll] Roll infra/recipes [presubmit_util] Use wait-on-cq subcommand
Instead of using the -wait flag on trigger-cq, use the new wait-on-cq
subcommand to 1) make the waiting explicit and 2) capture the pass or
fail status, which the new subcommand supports.

If we are respecting the CQ status (not collecting tryjobs explicitly),
then a fail status will result in a StepFailure.

Original-Bug: 102850
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/693146
Original-Revision: 70395fd4d680942c6fcf19ffc1886a2e130d6159
GitOrigin-RevId: 884b798a37cf670347fb27bb9c89e0dc901924d2
Change-Id: I66f6a8c1de1416919fa495a43dd5bd1b37a84287