[roll] Roll fuchsia [element_manager] Allow using Realm from parent
This CL changes the semantics of the `element_manager`
component to use the `fuchsia.component.Realm` protocol
from parent instead of from framework. This is done in order
to enable product variability of the capabilities routed to
the elements collection. Furthermore, the name of the collection
itself is now configurable using structured configuration.

In order to aid a soft transition, a component with the old
semantics and name will be kept until downstream users are moved
to using the new component.

Original-Bug: 95841
Original-Bug: 100933
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691451
Original-Revision: 531c37897d559bbb1d96a42e530030d64f01efc9
GitOrigin-RevId: fd6254a0d906c97247d3ada8905803e8e4e1565e
Change-Id: I7ddf2e327cf45105f694f2c4e73d3f0a201949db