[roll] Roll fuchsia [realm_builder][rust] allow local components to get stop notification
Adds a feature to rust realm builder with which local components can
register a stop notifier. Once this task is performed, any stop requests
that realm builder receives for the local component is reported over the
stop notifier, and it is then the local component's job to exit.

As with any component, if the component does not exit in a timely
fashion component manager will request the component be killed, which
causes realm builder to stop the local component immediately.

Original-Fixed: 82021
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692683
Original-Revision: c7ace08ef8d6c0ac5360bd52eb4e888f9fc06904
GitOrigin-RevId: 058eccf4dc06e9ad168d930a5f643533ee1de8f4
Change-Id: I6af481ded9a5d2dbc23079287040d36fe1db1b2c