[roll] Roll infra/recipes [static_checks] Check all files on integration changes
An integration change may contain a roll of the GN prebuilts that
changes the formatting rules in a way that results in existing files
being considered unformatted. So it's important to check formatting of
all files in this case to ensure the files are formatted before the roll
lands.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/693742
Original-Revision: 3a572eccb6d8b74a69171713374af9524c45f78c
GitOrigin-RevId: 0899b4479b0bac724cc883256e547a1fbe7089b8
Change-Id: I1fd3432b5978f797aab678968ddfa4bca1ddc7c1