[roll] Roll experiences [ermine] Migrate to new element_manager
This CL changes the element_manager component used
for launching apps to the newer one. Semantically,
the difference is that element_manager consumes
`fuchsia.component.Realm` from its parent component,
as opposed to the framework. This is done to allow
product variability in the capabilities routed to the
elements launched.

This CL is the first of two. This one changes the component
URL for element_manager to `...new_element_manager.cm` along
with the aforementioned changes. Once the older version of
element_manager is removed, the URL will be changed back
to `...element_manager.cm`.

Original-Bug: 95841
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/691391
Original-Revision: b1ea9d0b11b16fc959834ea5d6bad1ae1dc3c8fd
GitOrigin-RevId: d8d5f8f6a4df51ca89d3001bb37653dac02c09d0
Change-Id: I9763715523c77d3f73efcae4993ac0c36da22d75