[roll] Roll fuchsia [wlan][hw-sim] Send probe responses
Our mock AP is hardcoded as channel 1. Beacons are only sent when the hw-sim changes channel to the AP's channel (channel 1). Probe responses are never sent out, so we're wholly reliant on the beacons.

Usually, when a passive scan happens, the channel is tuned to channel 1 as part of the scan and then the beacon is sent.

It appears that sometimes, the Phy gets tuned to channel 1 prior to a scan. I think this happens when there was an association attempt earlier in the test. In those cases, the hw-sim is already on channel 1 when the scan starts, so no channel change happens, so no beacons are sent (and the test fails because we don't find an AP).

I tried to spelunk and find what  exactly causes the channel to be tuned, but it seems like something within SoftMAC and I hit a wall. It's not very important for the real world, it's only here in the hw-sim where the act of changing channels is the trigger for a beacon to be sent...

To fix this, we can change the hw-sim so that it responds to probe requests -- this means that when the hw-sim does an active scan, it'll find the AP, even if we were already on the same channel. There's probably other ways to fix it too, but this works in my testing.

Multiply: fuchsia-pkg://fuchsia.com/wlan-hw-sim-test#meta/reconnect-to-wpa2-network.cm: 200
Multiply: fuchsia-pkg://fuchsia.com/wlan-hw-sim-test#meta/connect-to-all-modern-bss-protections.cm: 200
Test: fx test wlan-hw-sim-test wlan-hw-sim-test-legacy-privacy
Original-Bug: 101944
Original-Bug: 101141
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689148
Original-Revision: a968249c4a3bae784875f4d10e403b3c223f960d
GitOrigin-RevId: dbb01ca4cbde6684a982c8729c722de77d5c8769
Change-Id: Ic399378c677498a7d2d48c072bd68f5d6c3ce53a