[roll] Roll fuchsia [diagnostics][persistence] v2 integration test
Port of the v1 Persistence integration test to v2. As part of this,

In Persistence:
 - After publishing the persisted data, Persistence adds a property
    "published" at inspect root with a value of monotonic now().
    This helps avoid race conditions from partially published data.
 - Now uses Structured Config

In SamplerTestController:
 - Make IncrementInt synchronous
 - Add SetOptional and RemoveOptional to support more test cases

In iquery:
 - Delete src/diagnostics/iquery/test/test_component/meta/test_component.cmx

PS1 of this CL is the same as fxr/691602 which flaked and was reverted.
PS2 of this CL fixes the race condition.

Multiply: fuchsia-pkg://fuchsia.com/persistence-integration-tests#meta/persistence-integration-test.cm
Original-Fixed: 84283
Original-Fixed: 102672
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693488
Original-Revision: 598a74dcd0a9c407a34c9f3ec45cb3f519f0dc64
GitOrigin-RevId: 61a0c012bf9914994d264c664894ad4057296369
Change-Id: Ic092aef3392860e09acc428bfafeca6037d65887