[roll] Roll fuchsia [zircon][utest] Do not check for the bottom 3 bits of DBGWVR
The core-address-tagging test fails on bringup.arm64-[thin]lto
builders in release mode. This is because an assert fails that checks
the bottom 3 bits of the DBGWVR that we assign. The initial reasoning
for this was because the bottom bits of DBGWVR are reserved according
to the spec. This is slightly incorrect though.

First is it's actually the bottom 2 bits that are reserved, not 3.
Second, these bottom 2 reserved bits are alsi ignored according to the
spec. From `D2.10.4 Watchpoint data address comparisons`:

```
DBGWVR<n>_EL1[1:0] are RES0 and are ignored.
```

Original-Fixed: 103228
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693834
Original-Revision: 872f5e5cae05c149295733f88654badc78d93467
GitOrigin-RevId: 47e667c041d779340142029155096fadaca6b6be
Change-Id: I054054de4ac06319b2f0415b7050b407436b2593