[roll] Roll fuchsia [audio_core] Route audio protocols directly from CFv2 audio_core
This is step 5 of the sequence described at fxbug.dev/90408#c20.

This CL also moves soundplayer out of src/sys/core/meta/core.cml since
soundplayer is not included in all products (such as recovery images).
This change is necessary in this CL because soundplayer depends on
audio_core, which is not in all products and therefore is not in
core.cml.

Original-Bug: 90408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665949
Original-Revision: 3de788a4fb6f95237f888c762e6d5dbd869b9c6c
GitOrigin-RevId: 29dc0a5ec4ccb3c119b2d9a55285dcca5e90e3f6
Change-Id: I8ba180e36cee36b141887c5bd0097ac9c14ef16a