[roll] Roll fuchsia [netstack3] Swap NS2/NS3 in images via GN flag
Add a flag to control which of two Fuchsia packages is used as the
system's network component. `fx set --args=use_netstack3=true ...`
builds with Netstack3 in place of Netstack2 as the system netstack. The
default behavior is to use Netstack2.

In addition, add missing services that the system netstack is required
to implement to Netstack3's manifest (even though Netstack3 does not
serve these yet) and a new Netstack3 component with the same name and
manifest URL as the Netstack2 component

Original-Bug: 100736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686314
Original-Revision: 89b07473d698b9e1a48ccd9747ce55192bd2f493
GitOrigin-RevId: a2a640e4542767eedc9dc7f1562aeb2f56f4179a
Change-Id: I5125991ed7c82fe94b64e7167b1a67e60fbf141e