[roll] Roll fuchsia [assembly] Establish contract for base-resolver config verifier
This change implements a simple golden file check for the new
base-resolver structured configuration. It includes a custom
implementation to deal with the need to have separate goldens for
"eng" and "non-eng" build types, and vacuously integrates the
implementation into assembly. The integration is vacuous because
it always compares a golden file to itself with TODOs in the
assembly build files to add the correct file paths when the
base-resolver structured configuration lands in fxrev.dev/687102.

Original-Bug: 102968
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694083
Original-Revision: dff369ec4bc7e59061a69d077f8d3fc515b37d15
GitOrigin-RevId: 44e2cc17cf57e2f889eb31b103054b01c6875da6
Change-Id: Ifba81c1cef50b0e96e510acb63ab601a1062f860