[roll] Roll fuchsia [kernel] Hide version string from compiler
kVersionString is filled in outside of the source file that defines
this variable. The compiler during lto picked up that because it is
const the size should always be 3, but this isn't actually the case.
By just hiding any provenance information about the return value of
version_string from the compiler we force it to actually call
strlen at runtime with the real value.

Original-Fixed: 103245
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693800
Original-Revision: 5c951495457e5ea53bf95773424723241cf8ca1a
GitOrigin-RevId: 3c602e07d874f6843a0ae9139c539ccce6343be3
Change-Id: I54fe77718614f8651fc85975a4ecca6ef220dfb5