[roll] Roll fuchsia [scenic][screen_capture] get buffer_count from importer
This CL updates the ScreenCaptureBufferCollectionImporter to map the
collection_id to the buffer count that will allow screen capture to
access the buffer_count instead of having the client import it with the args.

This CL also updated the renderer type from before where in would take
in a VkRenderer and instead now takes in any rederer type that inherits
Renderer. This allows for more flexibility.

Test: Tests ran on physical device and new tests added
$ fx test screen_capture_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691790
Original-Revision: 969337dfb8d1e4c476a0bf38f7951761a5e2849b
GitOrigin-RevId: c85aa0e0933363f52665e8fd0ca8b7b10181b4cc
Change-Id: I92b6e1f66c21ad156a4ec899d84a35effe89e657