[roll] Roll fuchsia [netstack3] Add UDP multicast subscription in Core
Add the ability to register for multicast membership via a socket
handle. The device used to register is chosen based on arguments passed
in and the device the socket is bound to, if any. If no device is
selected, the implementation will fail to register - the Linux-like
behavior of picking a device via the routing table will be implemented
in a future commit.

Original-Bug: 39479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690147
Original-Revision: c24d5b3ea7ed71c248d92dd5e013e49371ad4cff
GitOrigin-RevId: 8a225c482d8cd44df81628b28254df5443b0a34d
Change-Id: I213952db5e77204bb62fcf9ed4d3194a172f80bf