[roll] Roll fuchsia [kernel][vm] Plumb write flag on the VMO pin path
Pass down the IOMMU write permission to the VMO pinning code so that it
can distinguish between pinning for a read and a write. This is
essential to dirty track pages in the case of a pinned write. Once the
pages have been pinned and mapped in the IOMMU, the VM system will no
longer be able to trap the writes in order to mark pages dirty.

This commit just plumbs through the write flag but does not actually use
it yet. That will follow in a future CL. There is no change in behavior
with this commit.

Original-Bug: 63989
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693382
Original-Revision: ffe1835bdd076eb43b56d48e4032a429c1ba4780
GitOrigin-RevId: 1ab2a20bf5e1f6f92701897cd7f0a7b3feec30bd
Change-Id: I722161fb8b4549a78c356dfe83da8ca737cdd589