[roll] Roll fuchsia [fdf][ddk] Add driver_runtime_shared_lib sdk item
Change the driver_runtime implementation's shared_library to an sdk
shared library so that it can be consumed out of tree. This is needed
because in order to unit test drivers out of tree we need to link the
test binaries against symbols that are present in this shared library.

Original-Bug: 102050
Tests: No behavior changes so no tests are needed
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694443
Original-Revision: 25ad35be0c8588b874b04f1bc85a120b49a80aad
GitOrigin-RevId: b63819d60e554010b3ed3ad39ba795d60bc81a2b
Change-Id: I8ec57ed201f79c224ed4a80d377a556643dbbb61