[roll] Roll fuchsia [kernel][coresight] Do not constrain enumerated components
Previously we would error out if the components found in a ROM table
were not of "CoreSight class" or that of a particular legacy timestamp
generator. In this change, we remove that assertion, as such logic is an
overfit and does not leave room for leniency in encountering other
legacy components or vendor-specific wonkiness. A ROM table walker
should just be in the business of reporting its entries; validation of
expected components is more reasonably deferred to users.

Original-Bug: 103310
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694686
Original-Revision: 1ba67b63e9e38ccebbfde10e584cd2ce80f4f4cc
GitOrigin-RevId: bf99a5314788ddd57a05c9f7b19881f355f68359
Change-Id: I4c71a6e96ce4abf2966da448fe90710f827e42cf