[roll] Roll fuchsia [virtualization] Migrate in-process vsock tests
Migrate the existing in-process test suite to support the in development
out-of-process device.

Note that the new device doesn't support the full functionality of the
old device, so many test cases are behind an #if. Most of these won't
build, as they require updated helper functions and boilerplate.

Original-Bug: 97355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686153
Original-Revision: ec9145eebecce942d65b7483c72cc2d35d788d2f
GitOrigin-RevId: 6445875078147f9a8dd0a5f8929e77e9daf31b17
Change-Id: I5e178a5a0da433feb99ae4ba6bfc409aa7dfddb2