[roll] Roll fuchsia [settings] Add ManagedInspectMap utility
The existing InspectMap can be included in structs that implement
Inspect and adds all of its children to the inspect tree once when
attached. Our existing use cases have tended to insert and remove items
dynamically, which ends up requiring work to add the children to the
inspect hierarchy anyways. See the attached bug for more context.

This wrapper around HashMap automatically writes inserted items to the
inspect node that it manages. The existing InspectMap is renamed to
InspectWritableMap to better reflect its functionality, as items
inserted into it have to be manually written to inspect.

One usage of InspectWritableMap is migrated to ManagedInspectMap, along
with some test additions and cleanup. Most of the other usages of
InspectWritableMap can be migrated to ManagedInspectMap or simply
replaced with plain HashMaps.

Original-Bug: 103390
Test: fx test setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694515
Original-Revision: 0797ad72864c4e846394cddacd7b32859bdedf4d
GitOrigin-RevId: 395d46f4eaf1c2a02372e4ec9f1712775794c3c5
Change-Id: I7e6c2ba257ac679c548c1cf5f78b9da540228263