[roll] Roll fuchsia [chroot] Do not use argh for parsing arguments
chroot helpers are expected to forward all arguments to
the binary that they launch. This does not work correctly
for `--` flags because the argh crate attempts to parse
them as arguments of chroot.

Do manual parsing of the argument list instead and correctly
forward arguments to the launched binary.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695363
Original-Revision: 9a4e246627f61ef9cbc69c6f3446cc86f0933aaa
GitOrigin-RevId: 0db729389b506417f957500cff3a46f9f6d8417e
Change-Id: Ie5cd3bf3dc95aae836921cbd00d67e87ab42803e