[roll] Roll fuchsia [virtual_audio_util] Remove unused manifest
The fuchsia_shell_package rule defines command line tools. Ordinarily
these are run as ordinary executables with full permissions, not as a
component. Optionally, fuchsia_shell_package allows specifying a
manifest which enables the tool to be run as a component in addition to
as a command line tool.

virtual_audio_util is never run as a component, only a command line
tool, so remove the stale (unused) CFv1 manifest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695390
Original-Revision: 187d836360f3a2ada0165084c35db83a57ba3824
GitOrigin-RevId: 9b15b3d9a62d446b62c50b48e543416af54177a0
Change-Id: Ibd4a6052c43ab2eb017074e72fc1297f93eaf5bf