[roll] Roll fuchsia [component_manager] Forward flags for storage open
When a storage capability is opened, it is always opened
with RW flags, ignoring the flags that were passed in by
the client.

This is problematic if the client passes in the DESCRIBE
flag (which happens when `dash` attempts to `cd` into
a storage directory under `ffx component explore`)

Instead of cloning the storage directory, this change
reopens it using the client provided flags and mode.

Also removed the StartReason propagation for storage
capability. The reason is always the same (component is
accessing a storage capability)

Test: fx test component_manager_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694184
Original-Revision: fa83d8a01ffb2f43287db90878cf8af0de0038e8
GitOrigin-RevId: 9da1623212fe8f2f2fc0c1f0f38d94075459ecae
Change-Id: I20e9ae817ad017979820d6f45835b5d3041df88f