[roll] Roll fuchsia [settings] Convert light to Jobs
The light setting makes use of the custom change function feature of the
jobs framework, which allows using an arbitrary function to determine
whether or not to send an update to a listener rather than just using
equality. Custom change functions must be accompanied by an ID that is
used as part of the job signature, which is used by the jobs manager to
execute requests on a certain sequence and provide caches to store the
previous values for hanging gets.

This CL hashes the light group name to use as an ID so that subsequent
calls with the same light group name will hang until the value changes.

Original-Fixed: 79044
Test: fx test setui_service_tests
      manually verified on device that light API works
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682502
Original-Revision: dc57f210e77afb37a423736cb41b47169770a93f
GitOrigin-RevId: 5f3b996064cbdb95c1030055137e905007d07f2d
Change-Id: Iaf1464bb4dd8bdf725b27d2fbef83e94e21d865d