[roll] Roll fuchsia [microbenchmarks] Run fuchsia_microbenchmarks as a component
Previously it was being run as a shell command for the main "perf test
mode" invocation.

Temporarily resurrect the runTestComponent() method for Components v1
(renamed to runTestComponentV1() here), which was previously removed
in https://fuchsia-review.googlesource.com/c/fuchsia/+/671240.

Change both runTestComponentV1() and runTestComponentV2() to take a
processRuns argument for doing multiple process invocations, in order
to preserve the current characteristics of this test.  The latter
method is not exercised with processRuns > 1 yet, but that will be
exercised when this test is converted to use Components v2.

Original-Bug: 103374
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/microbenchmarks_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695107
Original-Revision: 6163008328235bd8f7445692b4389ec299766676
GitOrigin-RevId: 59eb0452ae32d1fcb5f8a590d95ef16798fa37d3
Change-Id: I9dadf3ac53cba19c4ba1a033155f156017474182