[roll] Roll fuchsia [identity] Delete Token Management.
Fuchsia previously supported acquiring authentication and identity
tokens such as OAuth AccessTokens from service providers such as Google.
We don't have any products that currently use or intend to use this
platform capability and we already unhooked it from the external API
surface.

This CL deletes the remaining library implementation, the two
AuthProvider components we were maintaining, and the remaining vestiges
inside the account system that consumed configuration for auth
providers. If needed in the future we could resurrect the auth providers
and a simplified version of the token management library to provide a
per-system-account service.

Its a bit of a sad day since Satsuki and I poured a good amount of
effort into this code, but less code is less maintenance.

Original-Bug: 101937,101939
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693908
Original-Revision: c455a54cff4c6d60afcbf123d6b4d2484cc1e4a9
GitOrigin-RevId: 542ff2c79e23f723100204e7fc9b6be2247573d6
Change-Id: I17f5d152da6c05b9c4e57c00a236c2c8e4ef4a48