[roll] Roll fuchsia [build] Show a hint when FIDL goldens are missing.
When users create new FIDL files, the goldens required by
fidl_api_compatibility_test are usually missing. The user
sees an unhelpful file not found error from Ninja.

This change provides a more helpful error and tells the
user how to create the file for the first time.

This involves using a depfile rather than action.inputs
to give the python script a chance to check the whether
the current and golden files exist and are non empty.

Previously we made it so that libraries with no declarations
available at the target API level would generate an empty FIDL
summary, preventing the change from appearing as though it
modified history. This change improves the situation even more:
It doesn't ask the user to create the golden file at all if the
current summary is empty.

Original-Bug: 101964
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694995
Original-Revision: 74de9eb27ceb6182bc4a260b243bc240585d7881
GitOrigin-RevId: 5a52f6fd616bb70f3dd634ffc5f7e4eed6627cd7
Change-Id: Icd2437a31b47e5c5a23f4856f98eae6d70380589