[roll] Roll fuchsia [nvme-cpp] New NVMe driver in C++
This CL is the first in a series of CLs that introduce a new, rewritten
NVMe driver to the tree.

This driver is written in C++ and uses more idiomatic code compared to
the older, C version (and it has tests!).

So far it does not support any I/O, but it sets up admin
submission/completion queues and runs IDENTIFY on the controller.

It also fills in some inspect information.

Test: fx test nvme-tests
Multiply: nvme-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686144
Original-Revision: d34341af41a20f8629609f0e4be051fb376b1c14
GitOrigin-RevId: 541864fb402b2aaf0ecf3a210330eb5a08b8efaa
Change-Id: I4e546106346c3622949a1ca4a04559287b8347e8