[roll] Roll fuchsia [virtualization][vsock] Reacquire signals before polling the socket
The default async socket behaviour is to cache the read/write signal as
an optimization, allowing a client to read or write until failure.

The vsock device never wants to fail a read, and isn't able to fail a
write due to the credit system. To address this, we must reacquire
these signals before polling the socket to ensure freshness.

Note that this codepath is relatively rare -- a write signal is only
polled when credit hits zero without any RX, and a read signal is only
polled when the socket is empty and the connection needs to wait anyway.

Original-Bug: 97355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695207
Original-Revision: 6d3d820f8c1a3c07db5e62eef1008fff52eabcfc
GitOrigin-RevId: 8b81c7ba69daa974de183d526f49759d28fd192d
Change-Id: I9b53d9ddd2c2d69b6a9bdbeae183fb41a11c820f