[roll] Roll fuchsia [starnix] Exit ffx starnix shell when remote task ends
The current code would not exit the shell because it waited for the
copy futures to end, which didn't happen until the user type something
on the terminal end generated an error trying to write to stdin.

This CL transforms the copy futures so that they never terminates, and
use try join so that the shell stops when the fuchsia component stops.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695343
Original-Revision: b0501cfde4c9e2ec76f4f97c1a1cde73501d1f89
GitOrigin-RevId: 6102aaa7ee8fdaeeb26de88e07d8aa58be8c67a9
Change-Id: I6f5079dd76db792d1f6fb417ec228598033fd2bd