[roll] Roll fuchsia [fdf][dfv2] Composite Device assembler
Create a class that will assemble composite devices in the way that
DFv1 expects them to be assembled.

This will help support device_add_composite in the compat shim.

Currently this class is not connected to anything, it will be
added to driver runner in a follow up CL.

Original-Bug: 103306
TEST: composite-assembler-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694808
Original-Revision: 52e30f0aa3d9e520127589dd3aba831155042416
GitOrigin-RevId: 4f45270821add336ed66bf57c1b5e0c07c368e6e
Change-Id: I953e914f1eaea3828606518dad0a6eed35bf25f1