[roll] Roll fuchsia [wlan][sme] Add Rust channel indirection with MLME
Change `create_sme` in the SME library to take in a Rust channel instead
of an MLME FIDL channel. With this CL, SME and MLME will communicate
with a combination of a bidirectional FIDL channel and two
uni-directional Rust channels (used for MLME requests and MLME events).
This adds an extra layer of indirection for now, but later on when we
unify SME and MLME, the intermediate FIDL channel can be removed and
SME and MLME will communicate through the two Rust channels directly.

Original-Bug: 66772
Test: Verify manually that connecting still works
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694062
Original-Revision: 2508ec6eff66946cd2751fbece47ded0cb0a13ae
GitOrigin-RevId: ae06a71a37b7a500164938523df364ff52988754
Change-Id: I2eacd685dc32f7b7478133941fdd54c78f14afff