[roll] Roll fuchsia [scenic][screen_capture2] implement Configure()
This CL works with the configure function of screen_capture2
which checks for argument existence and retreieves the buffer_count
from the ScreenCaptureBufferCollectionImporter.

Adds unittests that use the null_renderer as well as the
mock_renderer to ensure 100% testing. The mock_renderer is used to test
that if ImportBufferImage of the ScreenCaptureBufferCollectionImporter
fails, ReleaseBufferImage will be called the correct number of times.

Test: fx test screen_capture2_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694825
Original-Revision: 94157686183794b8b6168b8d2dec3540a4d96535
GitOrigin-RevId: cc55a4879e2dc88c91ace25aa5b4553a434d0994
Change-Id: I2ac55a1084522bfcc65a5abdb2c23f21f8820fd0