[roll] Roll fuchsia [fidl][cpp] unknown interactions server side
Change Natural Types FIDL servers to be able to handle Unknown
Interactions.

For Open and Ajar protocols, add an additional base-class,
fidl::UnknownInteractionHandler, which receives information about which
kind of unknown interaction was received and allows the server to
respond to it.

Change dispatching to also accept an UnknownInteractionHandlerEntry
which enables unknown interaction handling. The
UnknownInteractionHandlerEntry tells the Dispatch function whether to
handle only one-way or both one-way and two-way interactions and
provides a function to use to send the unknown interaction reply for the
transport. It also provides the method to dispatch the protocol
implementation's handle_unknown_interacton method.

Original-Bug: 88366
Test: fx test fidl_cpp_unknown_interactions_tests fidl_driver_transport_tests fidl_llcpp_unknown_interactions_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686963
Original-Revision: 8f69ccc6232762a25da904da4d3b634237d7ee61
GitOrigin-RevId: 482628c3745205ecd9fb187e1be8dff8c0899523
Change-Id: If30a9d5891c3aadf63b4f961af9ec18c7a2e28a6