[roll] Roll infra/recipes [fuchsia] Add shard_name resultdb tag
Flake Fetcher will use this tag to group tests by shard name. It
currently relies on the shard name in Sponge to avoid grouping
tefmocheck failures by environment. Tefmocheck failures are often
related to which tests are running in a shard, so if two different
shards produce a different result for a tefmocheck check then it doesn't
necessarily mean the failure mode is flaky, it could be deterministic
with respect to the set of tests in the shard.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/695983
Original-Revision: 713ba5a87954858f2ab46bc8347696a8baaad3db
GitOrigin-RevId: 997154dfe08f24ba0390196ffc254e40752301a6
Change-Id: I168814bc5989af7846048574cfa38c638712f2d3