[roll] Roll fuchsia [fx][rfc] Clarify "Area: FIDL" sections as samples
The RFC	generated contains "FIDL" area entries regardless of
actual area chosen. This can be	confusing. Expand the wording
of "Area: FIDL" headings to make it clearer that these are
sample and optional.

This is a continuation of fxr/693285 which was reverted by
fxr/695604. This CL updates a golden file that was affected but
presubmit tests didn't catch.

Test: fx test //tools/rfc
Original-Fixed: 103195
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695686
Original-Revision: 61a70c3f8e2b6775baa841ea253b9a4f994852a1
GitOrigin-RevId: 038988d581e3c6c1c7660ba1cad1577fcc611d3b
Change-Id: I4c12c0c55d8d69955f78c9112c1a7b5ec5322420