[roll] Roll fuchsia [sc] Rename procargs constant to match RFC.
Temporarily disables the fuchsia-element-tests which launch a
prebuilt component using structured config. The prebuilt depends on
the previous value for the procargs ordinal of 0x55, but we'd like
to use 0x1D to match the other VMO handles based on review feedback.

There are no OOT users of structured config yet, and this change appears
to pass LSC presubmits, so this should be safe. Future ABI-incompatible
changes to this protocol will need to be soft-migrated.

Original-Bug: 103694

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695223
Original-Revision: 6023ff0479b0b822e651ba4aa67d19fdd5df82aa
GitOrigin-RevId: da9f84254c21e78573ff01afa1b306fedadcad1c
Change-Id: Id7a9b141de5a8ea29fa55bd0df20f959f0daa8de