[roll] Roll fuchsia [overnet] Limit simultaneous ascendd connections.
This limits simultaneous ascendd connections to 5. This is based on
observations in testing from etryzelaar@google.com

This is intended as a stop-gap solution for fxbug.dev/87372 but will not
be a permanent or longer term fix.

Original-Bug: 87372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695967
Original-Revision: 9fb0fda791f34d03b261f8f451ababb749e5f2a5
GitOrigin-RevId: 596079b082f0e1d02bf2bf07833f3c24a6af74d6
Change-Id: Ie5e4b4f057a70c424abaabe5a171823d6b15ae99