[roll] Roll fuchsia [starnix][binder] Ease padding restriction on SG transactions
The length of a buffer does not need to be padded in the buffer data
structure sent to binder. What DOES need to be padded is the total
buffer size sent to binder, since the buffer start addresses need to be
aligned. This assumes that the length will be padded by binder instead.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696022
Original-Revision: b154782ae53a40f28395952b87fe32c9f32b76f5
GitOrigin-RevId: aacb09faba4f1557ffdbdc0096ebdd3908adb732
Change-Id: Ie34d3e61e49021daa48a4db27371b9550b4e9201