[roll] Roll fuchsia [driver_runtime] Synchronize irq callbacks.
Irqs are only supported for synchronized dispatchers,
and may only be unbound on the dispatcher thread on which
it was bound.

Added an AsyncIrq class that wraps around the user's
async_irq_t. It is similar to AsyncWait, except for the
fact there may be multiple callbacks per AsyncIrq.

Original-Bug: 99975

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689116
Original-Revision: 8c714413aafc956dd3416aa434c2997685078f30
GitOrigin-RevId: f4ac263080be8880aee6397cd78d3616f39584bc
Change-Id: I5757c8f04e9ba14249525e7c8b5767b15360449c