[roll] Roll fuchsia [fdf][driver-index] Update device group implementation
Rewrite how device groups are handled in the driver
index so that it can handle multibinding in DFv2.

Driver index now maps a list of device groups to
device group nodes. If it matches a node, it'll
return the list of device groups that contain that
node.

Original-Bug: 103206
Tests: driver-index-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694924
Original-Revision: 3e7f5edfbc5f794047576f483bcf5236deb6944c
GitOrigin-RevId: 075f75cb94e26dacaa81f7c85f7938478f527980
Change-Id: I7a47c1462354eaccd7331172c0ef50ecfe4a69af